### 2025-03-17 #7909 @editorialbot
[PRE REVIEW]: IntrinsicTimescales.jl: A Julia package to estimate intrinsic (neural) timescales (INTs) from time-series data
**Submitting author:** @duodenum96 (<a href="http://orcid.org/0000-0002-0942-899X">Yasir Çatal</a>)
**Repository:** https://github.com/duodenum96/IntrinsicTimescales.jl
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
