### 2025-03-17 #7909 @Kevin-Mattheus-Moerman
@duodenum96 - thanks for your submission to JOSS. We're currently managing a large backlog of submissions and the editors most appropriate for your area are all already rather busy.
