### 2025-03-19 54e8353
CompatHelper: bump compat for Turing to 0.37, (keep existing compat)
