+++ Project.toml
@@ -71,15 +71,15 @@ Plots = "1.40.9"
 Preferences = "1.4.3"
 ProgressMeter = "1.10.2"
 Random = "1.11.0"
-Revise = "3"
 Reexport = "1"
+Revise = "3"
 Romberg = "0.2.0"
 SciMLBase = "2.72.0"
 SciMLSensitivity = "7.72.0"
 SciMLStructures = "1.6.1"
 StaticArrays = "1.9.10"
-Turing = "0.36"
 Statistics = "1"
 StatsBase = "0.34"
 Test = "1.11"
+Turing = "0.36, 0.37"
 julia = "1.11"
