### 2025-03-28 e299d93
CompatHelper: bump compat for JET to 0.10, (keep existing compat)
