+++ Project.toml
@@ -57,7 +57,7 @@ DocumenterTools = "0.1.20"
 FFTW = "1"
 FastTransformsForwardDiff = "0.0.2"
 ForwardDiff = "0.10.38"
-JET = "0.9.12"
+JET = "0.9.12, 0.10"
 KernelDensity = "0.6.9"
 LinearAlgebra = "1.11"
 LogExpFunctions = "0.3.29"
@@ -71,15 +71,15 @@ Plots = "1.40.9"
 Preferences = "1.4.3"
 ProgressMeter = "1.10.2"
 Random = "1.11.0"
-Revise = "3"
 Reexport = "1"
+Revise = "3"
 Romberg = "0.2.0"
 SciMLBase = "2.72.0"
 SciMLSensitivity = "7.72.0"
 SciMLStructures = "1.6.1"
 StaticArrays = "1.9.10"
-Turing = "0.36"
 Statistics = "1"
 StatsBase = "0.34"
 Test = "1.11"
+Turing = "0.36"
 julia = "1.11"
