### 2025-04-04 #7909 @britta-wstnr
Hi @duodenum96!
I have already invited some potential reviewers (via email), but if you have any recommendations for reviewers, please let me know. You can mention them here in this thread, but **without** adding them, i.e., please do not use the @ sign in front of their handle. Some more information from our editorial bot in [this comment](https://github.com/openjournals/joss-reviews/issues/7909#issue-2924739180)!
Thanks a lot! 🙏
