### 2025-04-09 6fd5e99
fix bug when verbose=false in `pmc_abc`
