### 2025-04-09 e61c0df
remove unnecessary dependencies
