+++ Project.toml
@@ -17,7 +17,6 @@ DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 FastTransformsForwardDiff = "77fa7db0-1c81-401d-9fde-3592fc42b8bc"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
-JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
 KernelDensity = "5ab0869b-81aa-558d-bb23-cbf5423bbe9b"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 LogExpFunctions = "2ab3a3ac-af41-5b50-aa03-7779005ae688"
@@ -26,13 +25,11 @@ LombScargle = "fc60dff9-86e7-5f2f-a8a0-edeadbb75bd9"
 Missings = "e1d29d7a-bbdc-5cf2-9ac0-f12de2c33e28"
 NaNStatistics = "b946abbf-3ea7-4610-9019-9858bfdeaf2d"
 NonlinearSolve = "8913a72c-1f9b-4ce2-8d82-65094dcecaec"
-PkgSkeleton = "d254efa0-af53-535e-b7f1-03c1c9fbcbe7"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Preferences = "21216c6a-2e73-6563-6e65-726566657250"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
-Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 Romberg = "ff01fdde-158c-414c-948f-c94b4a236685"
 SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 SciMLSensitivity = "1ed8b502-d754-442c-8d5d-10ac956f44a1"
@@ -57,7 +54,6 @@ DocumenterTools = "0.1.20"
 FFTW = "1"
 FastTransformsForwardDiff = "0.0.2"
 ForwardDiff = "0.10.38"
-JET = "0.9.12"
 KernelDensity = "0.6.9"
 LinearAlgebra = "1.11"
 LogExpFunctions = "0.3.29"
@@ -66,20 +62,18 @@ LombScargle = "1.0.3"
 Missings = "1.2.0"
 NaNStatistics = "0.6.45"
 NonlinearSolve = "4"
-PkgSkeleton = "1.3.2"
 Plots = "1.40.9"
 Preferences = "1.4.3"
 ProgressMeter = "1.10.2"
 Random = "1.11.0"
-Revise = "3"
 Reexport = "1"
 Romberg = "0.2.0"
 SciMLBase = "2.72.0"
 SciMLSensitivity = "7.72.0"
 SciMLStructures = "1.6.1"
 StaticArrays = "1.9.10"
-Turing = "0.36"
 Statistics = "1"
 StatsBase = "0.34"
 Test = "1.11"
+Turing = "0.36"
 julia = "1.11"
