### 2025-04-09 0daee3d
add random seed for reproducibility of the tests
