+++ test/test_one_timescale.jl
@@ -6,9 +6,11 @@ using IntrinsicTimescales.OneTimescale
 using IntrinsicTimescales.Models
 using DifferentiationInterface
 using .ABC
+using Random
 
 
 @testset "OneTimescale Model Tests" begin
+    Random.seed!(123)
     # Setup test data and parameters
     dt = 0.01
     T = 100.0
