### 2025-04-17 40f7351
fix bug related to indexing of theta_star in draw_theta_pmc
