### 2025-04-18 26aad2b
cov_scale should not be a free parameter due to theory (fixed it)
