### 2025-04-24 b4ea7d4
- Add functions `residual_lorentzian_with_exponent!` and `lorentzian_with_exponent`. - Add new keyword argument `allow_variable_exponent` for `find_knee_frequency`. Default value is false. - Uniformize the output of `find_knee_frequency` so that it always returns sol.u [amplitude, knee, optionally exponent]. Previously it was returning [amplitude, knee] for vector inputs and only knee for array inputs. - Also new keyword argument `allow_variable_exponent` for `lorentzian_initial_guess`. If true, the initial guess for exponent is 2. - Corrections in documentation - Similar changes in `fooof_fit
[…]
