+++ src/utils/utils.jl
@@ -12,13 +12,16 @@ Module providing utility functions for time series analysis, including:
 """
 module Utils
 
+using Revise
 using Statistics
 using NonlinearSolve
 using Logging
 using Romberg
+
 export expdecayfit, find_oscillation_peak, find_knee_frequency, fooof_fit,
        lorentzian_initial_guess, lorentzian, expdecay, residual_expdecay!, fit_expdecay,
-       acw50, acw50_analytical, acw0, acweuler, tau_from_acw50, tau_from_knee, knee_from_tau, acw_romberg
+       acw50, acw50_analytical, acw0, acweuler, tau_from_acw50, tau_from_knee,
+       knee_from_tau, acw_romberg
 
 """
     expdecay(tau, lags)
@@ -78,7 +81,8 @@ function fit_expdecay(lags::AbstractVector{T}, acf::AbstractVector{T}) where {T 
     return sol.u[1]
 end
 
-function fit_expdecay(lags::AbstractVector{T}, acf::AbstractArray{T}; dims::Int=ndims(acf)) where {T <: Real}
+function fit_expdecay(lags::AbstractVector{T}, acf::AbstractArray{T};
+                      dims::Int=ndims(acf)) where {T <: Real}
     f = x -> fit_expdecay(lags, vec(x))
     return dropdims(mapslices(f, acf, dims=dims), dims=dims)
 end
@@ -105,11 +109,13 @@ Compute the ACW50 (autocorrelation width at 50%) along specified dimension.
 - Used for estimating characteristic timescales
 - Related to tau by: tau = -acw50/log(0.5)
 """
-function acw50(lags::AbstractVector{T}, acf::AbstractVector{T}; dims::Int=ndims(acf)) where {T <: Real}
+function acw50(lags::AbstractVector{T}, acf::AbstractVector{T};
+               dims::Int=ndims(acf)) where {T <: Real}
     lags[findfirst(acf .<= 0.5)]
 end
 
-function acw50(lags::AbstractVector{T}, acf::AbstractArray{T}; dims::Int=ndims(acf)) where {T <: Real}
+function acw50(lags::AbstractVector{T}, acf::AbstractArray{T};
+               dims::Int=ndims(acf)) where {T <: Real}
     f = x -> acw50(lags, vec(x))
     return dropdims(mapslices(f, acf, dims=dims), dims=dims)
 end
@@ -140,7 +146,8 @@ function acw0(lags::AbstractVector{T}, acf::AbstractVector{S}) where {T <: Real,
     end
 end
 
-function acw0(lags::AbstractVector{T}, acf::AbstractArray{S}; dims::Int=ndims(acf)) where {T <: Real, S <: Real}
+function acw0(lags::AbstractVector{T}, acf::AbstractArray{S};
+              dims::Int=ndims(acf)) where {T <: Real, S <: Real}
     f = x -> acw0(lags, vec(x))
     return dropdims(mapslices(f, acf, dims=dims), dims=dims)
 end
@@ -161,15 +168,16 @@ Compute the ACW at 1/e (≈ 0.368) along specified dimension.
 # Notes
 - For exponential decay, equals the timescale parameter tau
 """
-function acweuler(lags::AbstractVector{T}, acf::AbstractVector{S}) where {T <: Real, S <: Real}
-    lags[findfirst(acf .<= 1/ℯ)]
+function acweuler(lags::AbstractVector{T},
+                  acf::AbstractVector{S}) where {T <: Real, S <: Real}
+    lags[findfirst(acf .<= 1 / ℯ)]
 end
 
[…]
