### 2025-04-24 872d4c3
Update the documentation to reflect `allow_variable_exponent`
