+++ docs/src/practice/practice_4_psd.md
@@ -116,7 +116,7 @@ k = \frac{1}{2 \pi \tau} \\
 \tau = \frac{1}{2 \pi k}
 ```
 
-``\frac{A}{a^2 + \omega^2}`` is called a Lorentzian. The ``A`` term upstairs is just a normalization factor, not particularly interesting. ``\omega`` is our usual frequency. We also see that the timescale ``\tau`` is in there as well. This means that we need only to fit this Lorentzian to our power spectrum and we have access to timescale ``\tau``. To my knowledge, this technique was first introduced in [Chaudhuri et al., 2017](https://academic.oup.com/cercor/article/28/10/3610/4508768) and further developed in [Gao et al., 2020](https://elifesciences.org/articles/61277) to cover the cases where the exponent of ``omega`` downstairs is not exactly 2. Right now, IntrinsicTimescales.jl uses the simpler fitting that doesn't account for non-2 exponents. This should cover most cases. If you are getting funny results, I recommend the [FOOOF package](https://fooof-tools.github.io/fooof/) which uses the [Gao et al., 2020](https://elifesciences.org/articles/61277) technique. 
+``\frac{A}{a^2 + \omega^2}`` is called a Lorentzian. The ``A`` term upstairs is just a normalization factor, not particularly interesting. ``\omega`` is our usual frequency. We also see that the timescale ``\tau`` is in there as well. This means that we need only to fit this Lorentzian to our power spectrum and we have access to timescale ``\tau``. To my knowledge, this technique was first introduced in [Chaudhuri et al., 2017](https://academic.oup.com/cercor/article/28/10/3610/4508768) and further developed in [Gao et al., 2020](https://elifesciences.org/articles/61277) to cover the cases where the exponent of ``omega`` downstairs is not exactly 2. Since version 0.3.0, IntrinsicTimescales.jl supports both methods via the keyword  argument `allow_variable_exponent`. The default behavior for this is false, which sets the exponent to 2. This should cover most cases. If you are getting funny results, I recommend setting `allow_variable_exponent=true` (e.g. `acw(data, fs; acwtypes=:knee, allow_variable_exponent=true)`).
 
 Some interesting remarks. Consider very low frequencies: ``\omega \ll k``. When ``\omega`` is small, then ``\omega^2`` will be much smaller compared to ``k^2`` (think of if 2 is smaller than 3, then ``2^2=4`` is much smaller than ``3^2=9``). Then we can ignore the `\omega` term and our PSD reduces to ``\frac{A}{k^2}`` which is just a constant. Now consider big frequencies: ``\omega \gg k``. Since ``\omega`` is big, ``\omega^2`` is now much bigger than ``k`` and we can ignore ``k``. Then the power spectrum is ``\frac{A}{\omega^2}``. This is the so-called _scale-free_ power spectrum with a power-law exponent (PLE) of 2. If we take the logarithm, ``\log{\frac{A}{\omega^2}} \sim -2\omega`` meaning on the log-scale, we should see a straight line with a slope of 2. Right in between, there is a transition from a flat PSD to PLE=2 PSD. This is where ``k`` is approximately equal to ``\omega``. Alternatively, the frequency between the flat part and the PLE=2 part is the _knee frequency_ and also corresponds to your timescale up to a multiplicative constant. 
 
