### 2025-04-24 c566c24
fix typo in lorentzian_with_exponent formula
