### 2025-04-24 e3f5388
add allow_variable_exponent test in test_utils
