### 2025-04-24 39fda1d
fix wrong lorentzian formula for lorentzian_with_exponent
