### 2025-04-24 b7401be
add constrained option to utils
