+++ src/utils/utils.jl
@@ -12,10 +12,13 @@ Module providing utility functions for time series analysis, including:
 """
 module Utils
 
+using Revise
 using Statistics
 using NonlinearSolve
 using Logging
 using Romberg
+using Optimization
+using OptimizationOptimJL
 
 export expdecayfit, find_oscillation_peak, find_knee_frequency, fooof_fit,
        lorentzian_initial_guess, lorentzian, expdecay, residual_expdecay!, fit_expdecay,
@@ -232,7 +235,7 @@ Compute Lorentzian function values that allow variable exponent (PLE).
 - Vector of Lorentzian values: amp/(1 + (f/knee)²)
 """
 function lorentzian_with_exponent(f, u)
-    return u[1] ./ (1 .+ ( (f ./ u[2]) .^ u[3] ) )
+    return u[1] ./ (1 .+ ((f ./ u[2]) .^ u[3]))
 end
 
 # Define the residual function for NonlinearLeastSquares
@@ -241,9 +244,18 @@ function residual_lorentzian!(du, u, p)
     return nothing
 end
 
+# Optimization.jl doesn't like in-place functions
+function residual_lorentzian_constrained!(u, p)
+    return mean(sqrt.(abs2.(lorentzian(p[1], u) .- p[2])))
+end
+
 function residual_lorentzian_with_exponent!(du, u, p)
     du .= mean(sqrt.(abs2.(lorentzian_with_exponent(p[1], u) .- p[2])))
     return nothing
+end
+
+function residual_lorentzian_with_exponent_constrained!(u, p)
+    return mean(sqrt.(abs2.(lorentzian_with_exponent(p[1], u) .- p[2])))
 end
 
 """
@@ -268,7 +280,7 @@ Estimate initial parameters for Lorentzian fitting.
 """
 function lorentzian_initial_guess(psd::AbstractVector{<:Real},
                                   freqs::AbstractVector{<:Real}; min_freq::Real=freqs[1],
-                                  max_freq::Real=freqs[end], 
+                                  max_freq::Real=freqs[end],
                                   allow_variable_exponent::Bool=false)
     # Initial parameter guess
     # u[1]: estimate amplitude from low frequency power
@@ -283,6 +295,7 @@ function lorentzian_initial_guess(psd::AbstractVector{<:Real},
     else
         knee_guess = freqs[knee_guess_idx]
     end
+
     if allow_variable_exponent
         u0 = [initial_amp, knee_guess, 2.0]
     else
@@ -303,7 +316,7 @@ Find knee frequency by fitting Lorentzian to power spectral density.
 - `dims::Int=ndims(psd)`: Dimension along which to compute
 - `min_freq::T=freqs[1]`: Minimum frequency to consider
 - `max_freq::T=freqs[end]`: Maximum frequency to consider
-# Keyword arguments
+- `constrained::Bool=false`: Whether to use constrained optimization
 - `allow_variable_exponent::Bool=false`: Whether to allow variable exponent (PLE)
 
 # Returns
@@ -318,23 +331,49 @@ returns [amplitude, knee_frequency, exponent].
 function find_knee_frequency(psd::AbstractVector{T}, freqs::AbstractVector{T};
                              min_freq::T=freqs[1],
                              max_freq::T=freqs[end],
-                             allow_variable_exponent::Bool=false) where {T <: Real}
+                             allow_variable_exponent::Bool=false,
+                             constrained::Bool=false) where {T <: Real}
     # Initial parameter guess
     u0 = lorentzian_initial_guess(psd, freqs; min_freq=min_freq, max_freq=max_freq,
                                   allow_variable_exponent=allow_variable_exponent)
 
-    # Set up and solve the nonlinear least squares problem
     if allow_variable_exponent
[…]
