### 2025-04-24 8d48c02
fix tests, make tolerance stricter, remove noisy PSD test (was giving negative valued  PSDs)
