+++ test/test_utils.jl
@@ -5,6 +5,7 @@ using IntrinsicTimescales.OrnsteinUhlenbeck
 using IntrinsicTimescales.SummaryStats
 
 @testset "find_knee_frequency tests" begin
+    Random.seed!(666)
     # Test 1: Perfect Lorentzian
     @testset "Ideal Lorentzian" begin
         freqs = collect(0.1:0.1:100.0) / 1000.0
@@ -32,17 +33,17 @@ using IntrinsicTimescales.SummaryStats
 
     # Test 3: Variable exponent
     @testset "Variable exponent" begin
-        freqs = collect(0.0:0.1:100.0) / 1000.0
+        freqs = collect(0.1:0.1:100.0) / 1000.0
         f_knee = 10.0 / 1000.0
         amp = 100.0
         exponent = 1.5
         psd = @. amp / (1 + ((freqs/f_knee)^exponent))
-        amp, knee, exponent = fit_lorentzian(psd, freqs, allow_variable_exponent=true)
-        @test isapprox(knee, f_knee, rtol=0.2)
-        @test isapprox(exponent, 1.5, rtol=0.2)
+        psd2 = @. amp / (1 + ((freqs/f_knee)^2))
 
-        detected_knee = find_knee_frequency(psd, freqs, allow_variable_exponent=true)[2]
-        @test isapprox(detected_knee, f_knee, rtol=0.2)
+        estimated_amp, estimated_knee, estimated_exponent = find_knee_frequency(psd, freqs, allow_variable_exponent=true, constrained=true)
+        @test isapprox(estimated_knee, f_knee, rtol=0.01)
+        @test isapprox(estimated_exponent, exponent, rtol=0.01)
+
     end
 
     # 
@@ -57,8 +58,8 @@ using IntrinsicTimescales.SummaryStats
         knee1 = find_knee_frequency(psd, freqs, min_freq=0.1/1000.0, max_freq=50.0)[2]
         knee2 = find_knee_frequency(psd, freqs, min_freq=1.0 / 1000.0, max_freq=20.0 / 1000.0)[2]
         
-        @test isapprox(knee1, knee2, rtol=0.2)
-        @test isapprox(knee1, f_knee, rtol=0.2)
+        @test isapprox(knee1, knee2, rtol=0.01)
+        @test isapprox(knee1, f_knee, rtol=0.01)
     end
     
     # Test 4: Edge cases
@@ -86,9 +87,10 @@ using IntrinsicTimescales.SummaryStats
         num_trials = 100
         ou = generate_ou_process(tau, 1.0, dt, T, num_trials)
         psd, freqs = comp_psd(ou, 1/dt)
-        detected_knee = find_knee_frequency(psd, freqs, min_freq = freqs[1], max_freq = freqs[end])[2]
+        average_psd = mean(psd, dims=1)
+        detected_knee = find_knee_frequency(average_psd[:], freqs, min_freq = freqs[1], max_freq = freqs[end])[2]
         # Theoretically, tau = 1 / 2pi * f_knee
-        @test isapprox(tau, (1 / (2π * detected_knee)), rtol=2)
+        @test isapprox(tau, (1 ./ (2π .* detected_knee)), rtol=0.1)
 
         # Oscillation
         true_tau = 100.0
@@ -101,9 +103,10 @@ using IntrinsicTimescales.SummaryStats
         data = generate_ou_with_oscillation([true_tau, true_freq, true_coeff],
                                             dt, T, num_trials, 0.0, 1.0)
         psd, freqs = comp_psd(data, 1/dt)
-        detected_knee = find_knee_frequency(psd, freqs, min_freq = freqs[1], max_freq = freqs[end])[2]
+        average_psd = mean(psd, dims=1)
[…]
