### 2025-04-24 a3d5420
simpler tests for the fit function
