+++ test/test_one_timescale_and_osc_with_missing.jl
@@ -132,9 +132,9 @@ using Random
         # Custom parameters for faster testing
         param_dict = get_param_dict_abc()
         param_dict[:steps] = 3
-        param_dict[:max_iter] = 10000
+        param_dict[:max_iter] = 100
         param_dict[:distance_max] = 100.0
-        param_dict[:target_epsilon] = 1e-1
+        param_dict[:target_epsilon] = 10.0
         
         results = Models.fit(model, param_dict)
         samples = results.final_theta
