+++ test/test_one_timescale_with_missing.jl
@@ -166,8 +166,8 @@ using Random
             param_dict = get_param_dict_abc()
             param_dict[:steps] = 1
             param_dict[:distance_max] = 10.0
-            param_dict[:max_iter] = 10000
-            param_dict[:target_epsilon] = 1e-1
+            param_dict[:max_iter] = 100
+            param_dict[:target_epsilon] = 10.0
             
             results = Models.fit(model, param_dict)
             
