### 2025-04-24 dc58903
more lenient abc tests
