### 2025-04-24 910261d
more lenient lomb scargle test
