### 2025-04-24 65773e8
update dependencies (remove revise, add Optimization and OptimizationOptimJL
