+++ Project.toml
@@ -25,6 +25,8 @@ LombScargle = "fc60dff9-86e7-5f2f-a8a0-edeadbb75bd9"
 Missings = "e1d29d7a-bbdc-5cf2-9ac0-f12de2c33e28"
 NaNStatistics = "b946abbf-3ea7-4610-9019-9858bfdeaf2d"
 NonlinearSolve = "8913a72c-1f9b-4ce2-8d82-65094dcecaec"
+Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
+OptimizationOptimJL = "36348300-93cb-4f02-beb5-3c3902f8871e"
 Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 Preferences = "21216c6a-2e73-6563-6e65-726566657250"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
@@ -62,6 +64,8 @@ LombScargle = "1.0.3"
 Missings = "1.2.0"
 NaNStatistics = "0.6.45"
 NonlinearSolve = "4"
+Optimization = "4.1.2"
+OptimizationOptimJL = "0.4.2"
 Plots = "1.40.9"
 Preferences = "1.4.3"
 ProgressMeter = "1.10.2"
