### 2025-04-25 ecaf9f4
remove abc convergence tests
