+++ test/test_one_timescale.jl
@@ -172,7 +172,7 @@ using Random
             param_dict[:max_iter] = 100
             param_dict[:target_epsilon] = 10.0
             
-            results = Models.fit(model, param_dict)
+            results = int_fit(model, param_dict)
             
             # Test posterior properties
             @test results.MAP[1] isa Float64
@@ -203,7 +203,7 @@ using Random
             param_dict[:N] = 10000
             param_dict[:distance_max] = 500.0
             
-            results = Models.fit(model, param_dict)
+            results = int_fit(model, param_dict)
             
             # Test posterior properties
             @test results.MAP[1] ≈ true_tau atol=30.0
@@ -237,7 +237,7 @@ using Random
         #     param_dict[:N] = 10000
         #     param_dict[:distance_max] = 500.0
             
-        #     posterior_samples, posterior_MAP, abc_record = Models.fit(model, param_dict)
+        #     posterior_samples, posterior_MAP, abc_record = int_fit(model, param_dict)
             
         #     # Test posterior properties
         #     @test size(posterior_samples, 2) == 1
@@ -288,7 +288,7 @@ using Random
         )
 
         # Test with default parameters
-        result = IntrinsicTimescales.fit(model)
+        result = int_fit(model)
         samples = result.samples
         map_estimate = result.MAP
         posterior = result.variational_posterior
@@ -305,7 +305,7 @@ using Random
         param_dict[:n_elbo_samples] = 10
         param_dict[:autodiff] = AutoForwardDiff()
         
-        result2 = IntrinsicTimescales.fit(model, param_dict)
+        result2 = int_fit(model, param_dict)
         samples2 = result2.samples
         map_estimate2 = result2.MAP
         posterior2 = result2.variational_posterior
