### 2025-04-25 04fa9e6
rename fit to int_fit to avoid naming clash with other packages
