+++ src/models/one_timescale_and_osc.jl
@@ -11,7 +11,7 @@ using IntrinsicTimescales.Utils
 using IntrinsicTimescales
 using DifferentiationInterface
 
-export one_timescale_and_osc_model, OneTimescaleAndOscModel
+export one_timescale_and_osc_model, OneTimescaleAndOscModel, int_fit
 
 function informed_prior(data_sum_stats::Vector{<:Real}, lags_freqs; summary_method=:psd)
     if summary_method == :psd
@@ -379,7 +379,7 @@ function Models.distance_function(model::OneTimescaleAndOscModel, sum_stats, dat
 end
 
 """
-    Models.fit(model::OneTimescaleAndOscModel, param_dict=nothing)
+    int_fit(model::OneTimescaleAndOscModel, param_dict=nothing)
 
 Perform inference using the specified fitting method.
 
@@ -401,7 +401,7 @@ For ADVI method:
 - For ADVI: Uses Automatic Differentiation Variational Inference via Turing.jl
 - Parameter dictionary can be customized for each method (see get_param_dict_abc())
 """
-function Models.fit(model::OneTimescaleAndOscModel, param_dict::Dict=Dict())
+function Models.int_fit(model::OneTimescaleAndOscModel, param_dict::Dict=Dict())
     if model.fit_method == :abc
         if isempty(param_dict)
             param_dict = get_param_dict_abc()
