+++ src/models/one_timescale_and_osc_with_missing.jl
@@ -19,7 +19,7 @@ using IntrinsicTimescales
 using NaNStatistics
 using DifferentiationInterface
 
-export one_timescale_and_osc_with_missing_model, OneTimescaleAndOscWithMissingModel
+export one_timescale_and_osc_with_missing_model, OneTimescaleAndOscWithMissingModel, int_fit
 
 function informed_prior(data_sum_stats::Vector{<:Real}, lags_freqs; summary_method=:psd)
     if summary_method == :psd
@@ -355,7 +355,7 @@ function Models.distance_function(model::OneTimescaleAndOscWithMissingModel, sum
 end
 
 """
-    Models.fit(model::OneTimescaleAndOscWithMissingModel, param_dict=nothing)
+    int_fit(model::OneTimescaleAndOscWithMissingModel, param_dict=nothing)
 
 Perform inference using the specified fitting method.
 
@@ -377,7 +377,7 @@ For ADVI method:
 - For ADVI: Uses Automatic Differentiation Variational Inference via Turing.jl
 - Parameter dictionary can be customized for each method (see get_param_dict_abc())
 """
-function Models.fit(model::OneTimescaleAndOscWithMissingModel, param_dict::Dict=Dict())
+function Models.int_fit(model::OneTimescaleAndOscWithMissingModel, param_dict::Dict=Dict())
     if model.fit_method == :abc
         if isempty(param_dict)
             param_dict = get_param_dict_abc()
