+++ src/models/one_timescale.jl
@@ -15,7 +15,7 @@ using ComponentArrays
 using DifferentiationInterface
 
 
-export one_timescale_model, OneTimescaleModel
+export one_timescale_model, OneTimescaleModel, int_fit
 
 function informed_prior(data_sum_stats::Vector{<:Real}, lags_freqs; summary_method=:acf)
     if summary_method == :acf
@@ -315,7 +315,7 @@ function Models.distance_function(model::OneTimescaleModel, sum_stats, data_sum_
 end
 
 """
-    Models.fit(model::OneTimescaleModel, param_dict=nothing)
+    int_fit(model::OneTimescaleModel, param_dict=nothing)
 
 Perform inference using the specified fitting method.
 
@@ -337,7 +337,7 @@ For ADVI method:
 - For ADVI: Uses Automatic Differentiation Variational Inference via Turing.jl
 - Parameter dictionary can be customized for each method (see get_param_dict_abc())
 """
-function Models.fit(model::OneTimescaleModel, param_dict::Dict=Dict())
+function Models.int_fit(model::OneTimescaleModel, param_dict::Dict=Dict())
     if model.fit_method == :abc
         if isempty(param_dict)
             param_dict = get_param_dict_abc()
