+++ docs/src/practice/practice_4_psd.md
@@ -116,7 +116,7 @@ k = \frac{1}{2 \pi \tau} \\
 \tau = \frac{1}{2 \pi k}
 ```
 
-``\frac{A}{a^2 + \omega^2}`` is called a Lorentzian. The ``A`` term upstairs is just a normalization factor, not particularly interesting. ``\omega`` is our usual frequency. We also see that the timescale ``\tau`` is in there as well. This means that we need only to fit this Lorentzian to our power spectrum and we have access to timescale ``\tau``. To my knowledge, this technique was first introduced in [Chaudhuri et al., 2017](https://academic.oup.com/cercor/article/28/10/3610/4508768) and further developed in [Gao et al., 2020](https://elifesciences.org/articles/61277) to cover the cases where the exponent of ``omega`` downstairs is not exactly 2. Since version 0.3.0, IntrinsicTimescales.jl supports both methods via the keyword  argument `allow_variable_exponent`. The default behavior for this is false, which sets the exponent to 2. This should cover most cases. If you are getting funny results, I recommend setting `allow_variable_exponent=true` (e.g. `acw(data, fs; acwtypes=:knee, allow_variable_exponent=true)`). Yet another thing to note is the issue of constraints. In general, unconstrained problems, i.e. the problems where the parameters (for example the knee frequency) can take any possible value are solved more effectively. However, for hard problems, this can lead to the algorithm cheating by finding parameter combinations that are physically unrealistic (for example, negative frequencies for the knee frequency). The way to avoid this is by using the keyword argument `constrained`. When this is set to `true`, IntrinsicTimescales.jl switches the algorithm and the parameters are bounded in a realistic domain. See more details at [`acw`](@ref). 
+``\frac{A}{a^2 + \omega^2}`` is called a Lorentzian. The ``A`` term upstairs is just a normalization factor, not particularly interesting. ``\omega`` is our usual frequency. We also see that the timescale ``\tau`` is in there as well. This means that we need only to fit this Lorentzian to our power spectrum and we have access to timescale ``\tau``. To my knowledge, this technique was first introduced in [Chaudhuri et al., 2017](https://academic.oup.com/cercor/article/28/10/3610/4508768) and further developed in [Gao et al., 2020](https://elifesciences.org/articles/61277) to cover the cases where the exponent of ``\omega`` downstairs is not exactly 2. Since version 0.3.0, IntrinsicTimescales.jl supports both methods via the keyword  argument `allow_variable_exponent`. The default behavior for this is false, which sets the exponent to 2. This should cover most cases. If you are getting funny results, I recommend setting `allow_variable_exponent=true` (e.g. `acw(data, fs; acwtypes=:knee, allow_variable_exponent=true)`). Yet another thing to note is the issue of constraints. In general, unconstrained problems, i.e. the problems where the parameters (for example the knee frequency) can take any possible value are solved more effectively. However, for hard problems, this can lead to the algorithm cheating by finding parameter combinations that are physically unrealistic (for example, negative frequencies for the knee frequency). The way to avoid this is by using the keyword argument `constrained`. When this is set to `true`, IntrinsicTimescales.jl switches the algorithm and the parameters are bounded in a realistic domain. See more details at [`acw`](@ref). 
 
[…]
