### 2025-04-25 af787db
add development documentation badge to readme
