### 2025-04-25 ea8f860
fix function signature in the  documentation of fooof_fit
