### 2025-05-12 ccd65d6
CompatHelper: bump compat for DifferentiationInterface to 0.7, (keep existing compat)
