+++ Project.toml
@@ -48,7 +48,7 @@ BenchmarkTools = "1.6.0"
 ComponentArrays = "0.15.22"
 DSP = "0.8.0"
 DifferentialEquations = "7"
-DifferentiationInterface = "0.6.28"
+DifferentiationInterface = "0.6.28, 0.7"
 DifferentiationInterfaceTest = "0.9.1"
 Distributions = "0.25"
 Documenter = "1.8.0"
