### 2025-05-14 duodenum96/IntrinsicTimescales.jl#20 @behinger
[joss] dependencies
Hey!
I thought I'll already start with some review comments, even though the review didnt officially start. I have a bit of time and motivation now ;-)

## Project.toml / Dependencies
I think you could clean up your dependencies. There is things in there that should go to docs/Project.toml like `Documenter`, `BenchmarkTools`  or test/Project.toml (or [extras]) like JET. Also it is typically best practice to keep the dependent version flexible, e.g. `LombScargle = "1"` instead of `LombScargle = "1.0.3"`, as long as everyone follow SymVer we should be fine ;)

This will have the side-effect that installation will be much faster for users (I had to recompile 360 packages)

Please also think whether you can live with either Plots as an `PackageExtension` or use a Plot-Recipe to further reduce the dependencies, but that is a bit more involved.

Cheerio, Bene

PS: I read through the background and wow, very detailed and informativ!
