### 2025-05-14 duodenum96/IntrinsicTimescales.jl#20 @duodenum96
That's a very good point. I keep telling myself that I should handle that, loading the package takes an egregious amount of time. I'll do that as soon as possible.

Thanks for the compliments about the background, combined with the "Theory" section (which I still haven't got around to writing) I want that to be an entrypoint to INT research, especially for new grad students.

P.S. I was in your Unfold workshop the other day, very enlightening!
