### 2025-05-15 bdb437a
relaxed compatibility requirements
