+++ Project.toml
@@ -43,41 +43,41 @@ Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
 
 [compat]
-AbstractFFTs = "1.5.0"
-BenchmarkTools = "1.6.0"
-ComponentArrays = "0.15.22"
-DSP = "0.8.0"
+AbstractFFTs = "1"
+BenchmarkTools = "1"
+ComponentArrays = "0.15"
+DSP = "0.8"
 DifferentialEquations = "7"
-DifferentiationInterface = "0.6.28"
-DifferentiationInterfaceTest = "0.9.1"
+DifferentiationInterface = "0.6"
+DifferentiationInterfaceTest = "0.9"
 Distributions = "0.25"
-Documenter = "1.8.0"
-DocumenterTools = "0.1.20"
+Documenter = "1"
+DocumenterTools = "0.1"
 FFTW = "1"
 FastTransformsForwardDiff = "0.0.2"
-ForwardDiff = "0.10.38"
-KernelDensity = "0.6.9"
-LinearAlgebra = "1.11"
-LogExpFunctions = "0.3.29"
-Logging = "1.11.0"
-LombScargle = "1.0.3"
-Missings = "1.2.0"
-NaNStatistics = "0.6.45"
+ForwardDiff = "0.10"
+KernelDensity = "0.6"
+LinearAlgebra = "1"
+LogExpFunctions = "0.3"
+Logging = "1"
+LombScargle = "1"
+Missings = "1"
+NaNStatistics = "0.6"
 NonlinearSolve = "4"
-Optimization = "4.1.2"
-OptimizationOptimJL = "0.4.2"
-Plots = "1.40.9"
-Preferences = "1.4.3"
-ProgressMeter = "1.10.2"
-Random = "1.11.0"
+Optimization = "4"
+OptimizationOptimJL = "0.4"
+Plots = "1"
+Preferences = "1"
+ProgressMeter = "1"
+Random = "1"
 Reexport = "1"
-Romberg = "0.2.0"
-SciMLBase = "2.72.0"
-SciMLSensitivity = "7.72.0"
-SciMLStructures = "1.6.1"
-StaticArrays = "1.9.10"
+Romberg = "0.2"
+SciMLBase = "2"
+SciMLSensitivity = "7"
+SciMLStructures = "1"
+StaticArrays = "1"
 Statistics = "1"
 StatsBase = "0.34"
-Test = "1.11"
+Test = "1"
 Turing = "0.36"
 julia = "1.11"
