### 2025-05-15 1a7a762
replace DifferentialEquations.jl dependency with StochasticDiffEq dependency
