+++ Project.toml
@@ -39,6 +39,7 @@ SciMLStructures = "53ae85a6-f571-4167-b2af-e1d143709226"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 StatsBase = "2913bbd2-ae8a-5f71-8c99-4fb6c76f3a91"
+StochasticDiffEq = "789caeaf-c7a9-5a7d-9973-96adeb23e2a0"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 Turing = "fce5fe82-541a-59a6-adf8-730c64b5f9a0"
 
@@ -47,7 +48,6 @@ AbstractFFTs = "1"
 BenchmarkTools = "1"
 ComponentArrays = "0.15"
 DSP = "0.8"
-DifferentialEquations = "7"
 DifferentiationInterface = "0.6"
 DifferentiationInterfaceTest = "0.9"
 Distributions = "0.25"
@@ -78,6 +78,7 @@ SciMLStructures = "1"
 StaticArrays = "1"
 Statistics = "1"
 StatsBase = "0.34"
+StochasticDiffEq = "6.78.0"
 Test = "1"
 Turing = "0.36"
 julia = "1.11"
