### 2025-05-15 56cd347
relax dependency requirement for StochasticDiffEq
