+++ Project.toml
@@ -78,7 +78,7 @@ SciMLStructures = "1"
 StaticArrays = "1"
 Statistics = "1"
 StatsBase = "0.34"
-StochasticDiffEq = "6.78.0"
+StochasticDiffEq = "6"
 Test = "1"
 Turing = "0.36"
 julia = "1.11"
