### 2025-05-15 b329eef
Proper freqlims in lomb scargle tests
