+++ Project.toml
@@ -12,8 +12,6 @@ DifferentialEquations = "0c46a032-eb83-5123-abaf-570d42b7fbaa"
 DifferentiationInterface = "a0c0ee7d-e4b9-4e03-894e-1c5f64a51d63"
 DifferentiationInterfaceTest = "a82114a7-5aa3-49a8-9643-716bb13727a3"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
-Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
-DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 FastTransformsForwardDiff = "77fa7db0-1c81-401d-9fde-3592fc42b8bc"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
@@ -51,8 +49,6 @@ DSP = "0.8"
 DifferentiationInterface = "0.6"
 DifferentiationInterfaceTest = "0.9"
 Distributions = "0.25"
-Documenter = "1"
-DocumenterTools = "0.1"
 FFTW = "1"
 FastTransformsForwardDiff = "0.0.2"
 ForwardDiff = "0.10"
