### 2025-05-27 9440244
remove DifferentialEquations.jl dependency
