### 2025-05-27 5aef61b
correct a reference to an old function in `practice_5_bayesian.md`
