### 2025-05-27 f2c2ec4
move plots to an extension
