+++ docs/Project.toml
@@ -2,3 +2,6 @@
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 HypothesisTests = "09f84164-cd44-5f33-b23f-e6b0d136a0d5"
 IntrinsicTimescales = "10d2f5c9-08bb-4a99-a156-10b2ac3b4091"
+Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
+Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
