### 2025-05-27 58cc0c8
update documentation to indicate plotting is an extension
