+++ .github/workflows/CI.yml
@@ -1,9 +1,9 @@
 name: CI
 on:
-  push:
-    branches:
-      - master
-    tags: ['*']
+  # push:
+  #   branches:
+  #     - master
+  #   tags: ['*']
   pull_request:
   workflow_dispatch:
 concurrency:
