+++ docs/src/index.md
@@ -8,8 +8,8 @@ Pages = ["API.md"]
 Modules = [IntrinsicTimescales, IntrinsicTimescales.Models, IntrinsicTimescales.ABC, IntrinsicTimescales.ACW, IntrinsicTimescales.TuringBackend, IntrinsicTimescales.SummaryStats, 
     IntrinsicTimescales.Distances, IntrinsicTimescales.Utils, IntrinsicTimescales.OrnsteinUhlenbeck, IntrinsicTimescales.OneTimescale, 
     IntrinsicTimescales.OneTimescaleAndOsc, IntrinsicTimescales.OneTimescaleWithMissing, 
-    IntrinsicTimescales.OneTimescaleAndOscWithMissing, IntrinsicTimescales.Plotting]
-pages = ["IntrinsicTimescales.jl", ["core/model.jl", "core/one_timescale.jl", "core/one_timescale_and_osc.jl", "core/one_timescale_with_missing.jl", "core/one_timescale_and_osc_with_missing.jl"], "core/abc.jl", "core/turing_backend.jl", "stats/summary.jl", "stats/distances.jl", "utils/utils.jl", "utils/ou_process.jl", "core/plotting.jl"]
+    IntrinsicTimescales.OneTimescaleAndOscWithMissing, IntrinsicTimescales.IntPlottingExt]
+pages = ["IntrinsicTimescales.jl", ["core/model.jl", "core/one_timescale.jl", "core/one_timescale_and_osc.jl", "core/one_timescale_with_missing.jl", "core/one_timescale_and_osc_with_missing.jl"], "core/abc.jl", "core/turing_backend.jl", "stats/summary.jl", "stats/distances.jl", "utils/utils.jl", "utils/ou_process.jl", "core/plotting.jl", "ext/IntPlottingExt/IntPlottingExt.jl"]
 private = false
 ```
 
