### 2025-05-27 d42250d
working on documenting functions in plotting extension
