+++ src/core/plotting.jl
@@ -1,3 +1,15 @@
+module Plotting
 
+export acwplot, posterior_predictive
+
+"""
+Placeholder for acwplot in the extension package
+"""
 function acwplot end
+
+"""
+Placeholder for posterior_predictive in the extension package
+"""
 function posterior_predictive end
+
+end
