+++ Project.toml
@@ -5,11 +5,9 @@ version = "0.4.0"
 
 [deps]
 AbstractFFTs = "621f4979-c628-5d54-868e-fcf4e3e8185c"
-BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 ComponentArrays = "b0b7db55-cfe3-40fc-9ded-d10e2dbeff66"
 DSP = "717857b8-e6f2-59f4-9121-6e50c889abd2"
 DifferentiationInterface = "a0c0ee7d-e4b9-4e03-894e-1c5f64a51d63"
-DifferentiationInterfaceTest = "a82114a7-5aa3-49a8-9643-716bb13727a3"
 Distributions = "31c24e10-a181-5473-b8eb-7969acd0382f"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 FastTransformsForwardDiff = "77fa7db0-1c81-401d-9fde-3592fc42b8bc"
@@ -44,19 +42,11 @@ Plots = "91a5bcdd-55d7-5caf-9e0b-520d859cae80"
 [extensions]
 IntPlottingExt = "Plots"
 
-[extras]
-Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
-
-[targets]
-test = ["Test"]
-
 [compat]
 AbstractFFTs = "1"
-BenchmarkTools = "1"
 ComponentArrays = "0.15"
 DSP = "0.8"
 DifferentiationInterface = "0.6"
-DifferentiationInterfaceTest = "0.9"
 Distributions = "0.25"
 FFTW = "1"
 FastTransformsForwardDiff = "0.0.2"
@@ -87,3 +77,9 @@ StochasticDiffEq = "6"
 Test = "1"
 Turing = "0.36"
 julia = "1.11"
+
+[extras]
+Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+
+[targets]
+test = ["Test"]
