### 2025-05-27 9a8b6e1
Remove LogExpFunctions dependency from Project.toml
