+++ Project.toml
@@ -14,7 +14,6 @@ FastTransformsForwardDiff = "77fa7db0-1c81-401d-9fde-3592fc42b8bc"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
 KernelDensity = "5ab0869b-81aa-558d-bb23-cbf5423bbe9b"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
-LogExpFunctions = "2ab3a3ac-af41-5b50-aa03-7779005ae688"
 Logging = "56ddb016-857b-54e1-b83d-db4d58db5568"
 LombScargle = "fc60dff9-86e7-5f2f-a8a0-edeadbb75bd9"
 Missings = "e1d29d7a-bbdc-5cf2-9ac0-f12de2c33e28"
@@ -53,7 +52,6 @@ FastTransformsForwardDiff = "0.0.2"
 ForwardDiff = "0.10"
 KernelDensity = "0.6"
 LinearAlgebra = "1"
-LogExpFunctions = "0.3"
 Logging = "1"
 LombScargle = "1"
 Missings = "1"
