### 2025-05-27 duodenum96/IntrinsicTimescales.jl#20 @duodenum96
Hello again,

I just released version 0.4.1 of the package with the cleanup of the dependencies. Documenter is now in `docs/`, DocumenterTools, BenchmarkTools, DifferentiationInterfaceTest and LogExpFunctions are removed from dependencies. Test.jl is moved to extras. Also, plotting functions are now a package extension. As I mentioned above, DifferentialEquations.jl dependency is also dropped and replaced with StochasticDiffEq.jl. Together these should come with a significant decrease in precompilation and loading times.
