### 2025-05-28 #8261 @behinger
I started my review under [#20](https://github.com/duodenum96/IntrinsicTimescales.jl/issues/20)
