### 2025-05-28 duodenum96/IntrinsicTimescales.jl#22 @behinger
acw DocString
Given that `acw` is the main entry point, it might be helpful to make some implicits explicit (don't worry I wont check all docstrings :-P)

- `n_lags = nothing`, I dont know what happens in the default case - ah ok, it is written in the very botten in notes. Why not integrate the notes in the docstring on top? This is where I would have expected it
- return type is written `Vector` but it is actually  `ACWResults`

https://github.com/openjournals/joss-reviews/issues/8261
