### 2025-05-28 duodenum96/IntrinsicTimescales.jl#23 @behinger
[joss] Statement of Need
Could you copy the statement of need from the paper to your docs? It is one of the checkbox thingy; but I actually also think it would be helpful

--
edit: I would also include the cool graph you have in the paper in the docs. It's quite helpful
https://github.com/openjournals/joss-reviews/issues/8261
