### 2025-05-29 07c63fc
add reproducible results with random number generators for `ou_process` and appropriate tests
