+++ src/utils/ou_process.jl
@@ -7,6 +7,7 @@ Module for generating Ornstein-Uhlenbeck processes with various configurations.
 Uses DifferentialEquations.jl. 
 """
 module OrnsteinUhlenbeck
+using Revise
 using Random
 using Distributions: Distribution
 using Statistics
@@ -68,7 +69,7 @@ _prob_inplace = deq.SDEProblem(f_inplace, g_inplace, u0_inplace, (0.0, 1.0), p) 
 _prob_outofplace = deq.SDEProblem(f_outofplace, g_outofplace, u0_outofplace, (0.0, 1.0), p)
 
 """
-    generate_ou_process_sciml(tau, true_D, dt, duration, num_trials, standardize=true)
+    generate_ou_process_sciml(tau, true_D, dt, duration, num_trials, standardize=true; seed=nothing)
 
 Generate an Ornstein-Uhlenbeck process using DifferentialEquations.jl.
 
@@ -78,7 +79,10 @@ Generate an Ornstein-Uhlenbeck process using DifferentialEquations.jl.
 - `dt::Real`: Time step size
 - `duration::Real`: Total time length
 - `num_trials::Integer`: Number of trials/trajectories
-- `standardize::Bool=true`: Whether to standardize output
+- `standardize::Bool=true`: Whether to standardize output to match true_D
+- `rng::AbstractRNG=Random.default_rng()`: Random number generator for initial conditions
+- `deq_seed::Integer=nothing`: Random seed for DifferentialEquations.jl solver. If `nothing`, uses StochasticDiffEq.jl defaults. Note that for full replicability, 
+you need to set both `rng` and `deq_seed`. 
 
 # Returns
 - `Tuple{Matrix{Float64}, ODESolution}`: 
@@ -86,9 +90,22 @@ Generate an Ornstein-Uhlenbeck process using DifferentialEquations.jl.
   - Full SDE solution object
 
 # Notes
-- Uses SOSRA solver for efficiency
 - Switches between static and dynamic arrays based on num_trials
-- Standardizes output to match true_D if standardize=true
+
+Example: 
+```julia
+tau = 1.0
+true_D = 1.0
+dt = 0.01
+duration = 10.0
+num_trials = 100
+
+ou, _ = generate_ou_process_sciml(tau, true_D, dt, duration, num_trials, true)
+
+# Reproducible example
+deq_seed = 42
+ou, _ = generate_ou_process_sciml(tau, true_D, dt, duration, num_trials, true, rng=Xoshiro(42), deq_seed=deq_seed)
+```
 """
 function generate_ou_process_sciml(
     tau::Union{T, Vector{T}},
@@ -96,19 +113,27 @@ function generate_ou_process_sciml(
     dt::Real,
     duration::Real,
     num_trials::Integer,
-    standardize::Bool=true
+    standardize::Bool=true;
+    rng::AbstractRNG=Random.default_rng(),
+    deq_seed::Union{Integer, Nothing}=nothing
 ) where T <: Real
     
     p = [tau]
     if num_trials <= 20
-        u0 = SA[randn(num_trials)...] # Quick hack instead of ensemble problem
+        u0 = SA[randn(rng, num_trials)...] # Quick hack instead of ensemble problem
         prob = deq.remake(_prob_outofplace, p=p, u0=u0, tspan=(0.0, duration))
     else
-        u0 = randn(num_trials)
+        u0 = randn(rng, num_trials)
         prob = deq.remake(_prob_inplace, p=p, u0=u0, tspan=(0.0, duration))
     end
     times = dt:dt:duration
-    sol = deq.solve(prob, deq.SOSRA(); saveat=times, verbose=false)
+
[…]
