### 2025-05-29 c17336e
Extend the settings to `generate_ou_process` and `generate_ou_with_oscillation`
