+++ test/test_ou_process.jl
@@ -50,12 +50,10 @@ using Random
         T = 5.0
         num_trials = 5
         seed = 42
-        rng = Xoshiro(seed)
 
         # Generate two OU processes with the same seed
-        ou1 = generate_ou_process_sciml(tau, D, dt, T, num_trials, true; rng=rng, deq_seed=seed)[1]
-        rng = Xoshiro(seed)
-        ou2 = generate_ou_process_sciml(tau, D, dt, T, num_trials, true; rng=rng, deq_seed=seed)[1]
+        ou1 = generate_ou_process_sciml(tau, D, dt, T, num_trials, true; rng=Xoshiro(seed), deq_seed=seed)[1]
+        ou2 = generate_ou_process_sciml(tau, D, dt, T, num_trials, true; rng=Xoshiro(seed), deq_seed=seed)[1]
 
         # They should be identical
         @test ou1 ≈ ou2
@@ -72,6 +70,79 @@ using Random
         
         # These should likely be different (very small chance they're the same)
         @test !(ou4 ≈ ou5)
+    end
+
+    @testset "generate_ou_process seed functionality" begin
+        tau = 1.0
+        D = 1.0
+        dt = 0.01
+        T = 5.0
+        num_trials = 5
+        seed = 42
+
+        # Test reproducibility with both rng and deq_seed
+        ou1 = generate_ou_process(tau, D, dt, T, num_trials; rng=Xoshiro(seed), deq_seed=seed)
+        ou2 = generate_ou_process(tau, D, dt, T, num_trials; rng=Xoshiro(seed), deq_seed=seed)
+
+        # They should be identical
+        @test ou1 ≈ ou2
+
+        # Test with different deq_seed only
+        ou3 = generate_ou_process(tau, D, dt, T, num_trials; rng=Xoshiro(seed), deq_seed=seed+1)
+        @test !(ou1 ≈ ou3)
+
+        # Test with different rng only
+        ou4 = generate_ou_process(tau, D, dt, T, num_trials; rng=Xoshiro(seed+1), deq_seed=seed)
+        @test !(ou1 ≈ ou4)
+
+        # Test default behavior (no seeds)
+        ou5 = generate_ou_process(tau, D, dt, T, num_trials)
+        ou6 = generate_ou_process(tau, D, dt, T, num_trials)
+        @test !(ou5 ≈ ou6)  # Should be different
+    end
+
+    @testset "generate_ou_with_oscillation seed functionality" begin
+        theta = [1.0, 0.5, 0.7]  # [tau, frequency, coefficient]
+        dt = 0.01
+        T = 5.0
+        num_trials = 5
+        data_mean = 0.0
+        data_sd = 1.0
+        seed = 42
+
+        # Test reproducibility with both rng and deq_seed
+        osc1 = generate_ou_with_oscillation(theta, dt, T, num_trials, data_mean, data_sd; 
+                                          rng=Xoshiro(seed), deq_seed=seed)
+        osc2 = generate_ou_with_oscillation(theta, dt, T, num_trials, data_mean, data_sd; 
+                                          rng=Xoshiro(seed), deq_seed=seed)
+
+        # They should be identical
+        @test osc1 ≈ osc2
+
+        # Test with different deq_seed
+        osc3 = generate_ou_with_oscillation(theta, dt, T, num_trials, data_mean, data_sd; 
+                                          rng=Xoshiro(seed), deq_seed=seed+1)
+        @test !(osc1 ≈ osc3)
+
+        # Test with different rng
+        osc4 = generate_ou_with_oscillation(theta, dt, T, num_trials, data_mean, data_sd; 
[…]
