### 2025-05-29 2874ccd
update docstrings of abc.jl
