+++ src/core/abc.jl
@@ -17,7 +17,7 @@ using KernelDensity
 export basic_abc, pmc_abc, effective_sample_size, weighted_covar, find_MAP, get_param_dict_abc, 
     abc_results, ABCResults
 
-    """
+"""
     ABCResults
 
 Container for ABC results to standardize plotting interface.
@@ -29,6 +29,7 @@ Container for ABC results to standardize plotting interface.
 - `weights_history::Vector{Vector{Float64}}`: History of weights
 - `final_theta::Matrix{Float64}`: Final accepted parameter values
 - `final_weights::Vector{Float64}`: Final weights
+- `MAP::Vector{Float64}`: Maximum A Posteriori (MAP) estimate of parameters
 """
 struct ABCResults
     theta_history::Vector{Matrix{Float64}}
@@ -43,7 +44,19 @@ end
 """
     abc_results(output_record::Vector{NamedTuple})
 
-Construct abc_results from PMC-ABC output record.
+Construct an `ABCResults` struct from PMC-ABC output records.
+
+# Arguments
+- `output_record::Vector{NamedTuple}`: Vector of named tuples containing PMC-ABC iteration results. 
+  Each tuple must contain:
+    - `theta_accepted`: Accepted parameter values
+    - `epsilon`: Epsilon threshold value
+    - `n_accepted`: Number of accepted samples
+    - `n_total`: Total number of samples
+    - `weights`: Importance weights
+
+# Returns
+- `ABCResults`: Struct to contain ABC results. See the documentation for `ABCResults` for more details.
 """
 function abc_results(output_record::Vector{NamedTuple})
     n_steps = length(output_record)
@@ -106,31 +119,39 @@ end
 """
     basic_abc(model::Models.AbstractTimescaleModel; kwargs...)
 
-Perform basic ABC rejection sampling.
+Perform basic ABC rejection sampling. The algorithm stops either when `max_iter` is reached or 
+when `min_accepted` samples have been accepted.
 
 # Arguments
 - `model::Models.AbstractTimescaleModel`: Model to perform inference on
-- `epsilon::Float64`: Acceptance threshold
-- `max_iter::Integer`: Maximum number of iterations
-- `min_accepted::Integer`: Minimum number of accepted samples required
-- `pmc_mode::Bool=false`: Whether to use PMC proposal distribution
-- `weights=Array{Float64}`: Importance weights (used in PMC mode)
-- `theta_prev=Array{Float64}`: Previous parameters (used in PMC mode)
-- `tau_squared=Array{Float64}`: Covariance matrix (used in PMC mode)
-- `show_progress::Bool=true`: Whether to show progress bar
+- `epsilon::Float64`: Acceptance threshold for distance between simulated and observed data
+- `max_iter::Integer`: Maximum number of iterations to perform
+- `min_accepted::Integer`: Minimum number of accepted samples required before stopping
+- `pmc_mode::Bool=false`: Whether to use PMC proposal distribution instead of prior
+- `weights::Array{Float64}`: Importance weights for PMC sampling (only used if pmc_mode=true)
+- `theta_prev::Array{Float64}`: Previous parameters for PMC sampling (only used if pmc_mode=true)
+- `tau_squared::Array{Float64}`: Covariance matrix for PMC sampling (only used if pmc_mode=true)
+- `show_progress::Bool=true`: Whether to show progress bar with acceptance count and speed
 
 # Returns
 NamedTuple containing:
-- `samples`: All proposed parameters
-- `isaccepted`: Boolean mask of accepted samples
-- `theta_accepted`: Accepted parameters
-- `distances`: Distances for all proposals
-- `n_accepted`: Number of accepted samples
-- `n_total`: Total number of iterations
[…]
