### 2025-05-29 9724dfd
Update docstrings for `acw.jl`
