### 2025-05-29 d38734d
update docstrings for `turing_backend.jl`
