### 2025-05-29 4cfed93
update docstrings of one_timescale_and_osc_with_missing.jl
