### 2025-05-29 051df92
update docstrings for `one_timescale_and_osc_model`
