### 2025-05-29 3b33c78
update one_timescale.jl docstrings
