### 2025-05-29 954e0c7
update docstring for distances.jl
