### 2025-05-29 b0cfff2
update docstrings for `ou_process.jl`
