### 2025-05-29 3389728
update docstring for `acw.jl` to include detailed fields of the ACWResults structure
