### 2025-05-29 6310346
set size_threshold to nothing in `makedocs`
