### 2025-05-29 86faffa
move  `size_threshold=nothing` to `format=(...)`
