### 2025-05-30 67ca716
Add cool diagram and statement of need to the homepage
