### 2025-05-30 f9bc71d
fix the syntax for cool diagram to match Documenter.jl style
