### 2025-05-30 dd2f17e
Add 3-parameter exponential decay function and fitting method to `utils.jl`.
