### 2025-05-30 d71cba5
Enhance `acw` function by adding `skip_zero_lag` parameter
