### 2025-05-30 bfebee7
add tests for `skip_zero_lag`, fix bugs in the functions expdecay_3_parameters and residual_expdecay_3_parameters!
