### 2025-05-30 duodenum96/IntrinsicTimescales.jl#22 @behinger
nice!!
