### 2025-06-02 1958f81
return NaN if ACW-50 and ACW-e are not found
