### 2025-06-02 1d764e1
return NaN if ACW-50 or ACW-e is not found
