### 2025-06-02 5c4f0de
some refactoring in `practice_3_ou.md`
