### 2025-06-03 394c698
Embarrasingly forgot to add the `constrained` argument to `acw.jl` introduced in v0.3.0. Fixing it here along with some docstring fixes
