### 2025-06-03 00fa9aa
talk about the new tutorial in `home.md`
