### 2025-06-03 239ade2
add tutorial navigating the forest of INT metrics
