+++ docs/src/tutorial/tutorial_1_acw.md
@@ -1,0 +1,99 @@
+# Navigating the Forest of INT Metrics
+
+The main user-facing function in IntrinsicTimescales.jl is [`acw`](../acw.md). For reference, see the section [Model-Free Timescale Estimation](../acw.md). `acw` offers quite a number of different ways to estimate timescales. The situation might be confusing for newcomers and even seasoned researchers. INTs had been a part of my life for the last 5 years and even I occasionally find myself unsure on what to do. As experimental researchers, we do not have the luxury of a ground truth that we can use to validate our estimations. If we had, we would not need to estimate anything in the first place. The best we can do is to compare our results to the simulations where the ground truth is known, literature and theory. I believe that experimental researchers are entitled to look at computational researchers and package developers and ask them "Okay, this is all well and good but what I _should_ do?". The problem is, there is almost never a gold standard / best practice that we can use and get the best results. Therefore I opted to avoid recommending one. Nonetheless, there are certain rules of thumb and certain things to _not_ do. The purpose of this section is to talk about them. 
+
+I should note that this section will not go into detailed discussions about each metric. This is done in the [Practice](../practice/practice_intro.md) section. If you are not familiar with the INTs, I would highly recommend you to go through it to build your intuition slowly. The timescales of understanding are unfortunately slow :upside_down_face:. 
+
+In the function `acw`, one can set the different methods to use as simply:
+
+```julia
+acwresults = acw(data, fs; acwtypes=[:acw0, :acw50, :acweuler, :auc, :tau, :knee])
+```
+
+Different metrics specified in `acwtypes` is automatically calculated and returned in `acwresults`. Below, I'll try to show you the mental workflow on deciding which metric to use. 
+
+## Getting ACW-0 Out of the Way:
+
[…]
