### 2025-06-03 8d5c0cd
finish the tutorial of `Navigating the Forest of INT Metrics`
