+++ docs/src/tutorial/tutorial_1_acw.md
@@ -14,7 +14,7 @@ Different metrics specified in `acwtypes` is automatically calculated and return
 
 ## Getting ACW-0 Out of the Way:
 
-Never use `:acw0` in serious research. I considered to not include it in the final package but I know that some people will ask for it anyway. So I decided to put it here but warn against it. There are serious problems with ACW-0. Consider when an exponential decay function $f(x) = e^{\lambda x}$ reaches zero. There are two cases here. In the first case $\lambda$ is a real number. This is the pure exponential decay function without oscillations we usually encounter in fMRI or calcium imaging (after averaging over channels). In this case, $f(x)$ never touches zero. You might protest this suggestion by saying "but look, in my data I plot the ACF and it clearly touches zero". This is due to finite data bias. In all data, you have this bias and ACW-0 amplifies it. Now let's consider the second case where $\lambda$ is complex. This is the case we deal with usually in EEG / MEG data with oscillatorions. Let $ \lambda = a + bi $, then by Euler's formula, we have $ f(x) = e^{ax} e^{ibx} = e^{ax} (cos(bx) + i sin (bx) )$. Since our data is real valued, the actual autocorrelation function is the real part of this. Here, the real part of $\lambda$, $\mathscr{R}(\lambda) = a$ is the decay rate of the ACF and therefore, the timescale. Now we can ask ourselves when the ACF touches zero. Solving $cos(bx) = 0$ for x, we will get $x = \frac{\pi}{2b} $. Clearly ACW-0 here does not show the timescale, it just shows the oscillatory artifact. If you are interested in oscillations, that is perfectly fine but there are way better tools for that, such as the Fourier transform. This is the case when you have long data. In the case of smaller data, finite data bias might beat oscillatory artifacts. Now there is one more case I haven't talked about: multiple oscillations in ACF. But it is clear that that will suffer from the same problem. I used to think ACW-0 is justified in data with very low sampling rate, such as fMRI, but area under the curve (`:auc`) and estimating the decay rate directly (`:tau`) especially with the option `skip_first_lag = true` are much better methods. 
[…]
