### 2025-06-03 e5842eb
refer to citations in home, clarify tutorial, add these to `make.jl`
