### 2025-06-03 288f144
add `contributing.md`
