+++ docs/src/contributing.md
@@ -1,0 +1,29 @@
+# [Contributing guidelines](@id contributing)
+
+First of all, thanks for the interest!
+
+We welcome all kinds of contribution, including, but not limited to code, documentation, examples, configuration, issue creating, etc. Every minute of an open source developer's time is considered the most valuable gift: their time and appreciated. 
+
+Needless to say, be polite and respectful when interacting with strangers while you are contributing. Currently, IntrinsicTimescales.jl is maintained by only one person, it would be great to have other developers / contributers. Therefore please be patient if I can't address your problems in a timely manner. 
+
+## Bug reports and discussions
+
+If you think you found a bug, feel free to open an [issue](https://github.com/duodenum96/IntrinsicTimescales.jl/issues).
+Focused suggestions and requests can also be opened as issues.
+Before opening a pull request, start an issue or a discussion on the topic, please.
+
+## Working on an issue
+
+If you found an issue that interests you, comment on that issue what your plans are.
+If the solution to the issue is clear, you can immediately create a pull request.
+Otherwise, say what your proposed solution is and wait for a discussion around it.
+
+!!! tip
+    Feel free to ping us after a few days if there are no responses.
+
+If your solution involves code (or something that requires running the package locally), check the [developer documentation](developer.md).
+Otherwise, you can use the GitHub interface directly to create your pull request.
+
+## Discussions
+
+IntrinsicTimescales.jl has a [discussions](https://github.com/duodenum96/IntrinsicTimescales.jl/discussions) page on github which is intended for more relaxed discussions around the package and INTs in general. Feel free to open free-form discussions there. 
