### 2025-06-03 8fe44d8
refer to contributing guidelines in homepage, add contributions and developer docs to `make.jl`.
