### 2025-06-03 f7109d9
fix bug in make.jl, add figure for navigating... tutorial
