### 2025-06-03 8fe4d94
bump version to 0.5.0
