### 2025-06-05 044631e
CompatHelper: bump compat for Turing to 0.39, (keep existing compat)
