### 2025-06-16 4868a24
fix issues 31 and 32
