+++ Project.toml
@@ -19,12 +19,14 @@ LombScargle = "fc60dff9-86e7-5f2f-a8a0-edeadbb75bd9"
 Missings = "e1d29d7a-bbdc-5cf2-9ac0-f12de2c33e28"
 NaNStatistics = "b946abbf-3ea7-4610-9019-9858bfdeaf2d"
 NonlinearSolve = "8913a72c-1f9b-4ce2-8d82-65094dcecaec"
+OhMyThreads = "67456a42-1dca-4109-a031-0a68de7e3ad5"
 Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
 OptimizationOptimJL = "36348300-93cb-4f02-beb5-3c3902f8871e"
 Preferences = "21216c6a-2e73-6563-6e65-726566657250"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
+Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 Romberg = "ff01fdde-158c-414c-948f-c94b4a236685"
 SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 SciMLSensitivity = "1ed8b502-d754-442c-8d5d-10ac956f44a1"
@@ -57,6 +59,7 @@ LombScargle = "1"
 Missings = "1"
 NaNStatistics = "0.6"
 NonlinearSolve = "4"
+OhMyThreads = "0.8.3"
 Optimization = "4"
 OptimizationOptimJL = "0.4"
 Plots = "1"
@@ -64,6 +67,7 @@ Preferences = "1"
 ProgressMeter = "1"
 Random = "1"
 Reexport = "1"
+Revise = "3.8.0"
 Romberg = "0.2"
 SciMLBase = "2"
 SciMLSensitivity = "7"
