### 2025-06-18 bcf8752
add parallel kwarg and OhMyThreads based parallelism to all the main functions used in `acw`
