+++ src/utils/utils.jl
@@ -12,12 +12,14 @@ Module providing utility functions for time series analysis, including:
 """
 module Utils
 
+using Revise
 using Statistics
 using NonlinearSolve
 using Logging
 using Romberg
 using Optimization
 using OptimizationOptimJL
+using OhMyThreads
 
 export expdecayfit, find_oscillation_peak, find_knee_frequency, fooof_fit,
        lorentzian_initial_guess, lorentzian, expdecay, residual_expdecay!, fit_expdecay,
@@ -73,7 +75,7 @@ This is used in `acw` with the setting `skip_zero_lag=true`.
 - Vector of A*(exp(-t/tau) + B) values
 """
 function expdecay_3_parameters(p, lags)
-    return p[1]*(exp.(- (lags ./ p[2]) ) .+ p[3])
+    return p[1] * (exp.(-(lags ./ p[2])) .+ p[3])
 end
 
 function residual_expdecay_3_parameters!(du, u, p)
@@ -108,9 +110,14 @@ function fit_expdecay(lags::AbstractVector{T}, acf::AbstractVector{T}) where {T 
 end
 
 function fit_expdecay(lags::AbstractVector{T}, acf::AbstractArray{T};
-                      dims::Int=ndims(acf)) where {T <: Real}
+                      dims::Int=ndims(acf); parallel::Bool=false) where {T <: Real}
+    slices = get_slices(acf, dims=dims)
     f = x -> fit_expdecay(lags, vec(x))
-    return dropdims(mapslices(f, acf, dims=dims), dims=dims)
+    if parallel
+        return tmap(f, slices)
+    else
+        return map(f, slices)
+    end
 end
 
 """
@@ -129,7 +136,8 @@ Excludes lag 0 from fitting.
 # Notes
 - Initial guess: A=0.5, tau from ACW50, B=0.0
 """
-function fit_expdecay_3_parameters(lags::AbstractVector{T}, acf::AbstractVector{T}) where {T <: Real}
+function fit_expdecay_3_parameters(lags::AbstractVector{T},
+                                   acf::AbstractVector{T}) where {T <: Real}
     u0 = [0.5, tau_from_acw50(acw50(lags, acf)), 0.0]
     prob = NonlinearLeastSquaresProblem(NonlinearFunction(residual_expdecay_3_parameters!,
                                                           resid_prototype=zeros(1)), u0,
@@ -139,9 +147,14 @@ function fit_expdecay_3_parameters(lags::AbstractVector{T}, acf::AbstractVector{
 end
 
 function fit_expdecay_3_parameters(lags::AbstractVector{T}, acf::AbstractArray{T};
-                                   dims::Int=ndims(acf)) where {T <: Real}
+                                   dims::Int=ndims(acf); parallel=false) where {T <: Real}
+    slices = get_slices(acf, dims=dims)
     f = x -> fit_expdecay_3_parameters(lags, vec(x))
-    return dropdims(mapslices(f, acf, dims=dims), dims=dims)
+    if parallel
+        return tmap(f, slices)
+    else
+        return map(f, slices)
+    end
 end
 
 acw50_analytical(tau) = -tau * log(0.5)
@@ -177,9 +190,14 @@ function acw50(lags::AbstractVector{T}, acf::AbstractVector{T};
 end
 
 function acw50(lags::AbstractVector{T}, acf::AbstractArray{T};
-               dims::Int=ndims(acf)) where {T <: Real}
+               dims::Int=ndims(acf), parallel::Bool=false) where {T <: Real}
+    slices = get_slices(acf, dims=dims)
     f = x -> acw50(lags, vec(x))
[…]
