### 2025-06-19 b24b5f2
proper handling of dimensions
