+++ src/utils/utils.jl
@@ -110,13 +110,13 @@ function fit_expdecay(lags::AbstractVector{T}, acf::AbstractVector{T}) where {T 
 end
 
 function fit_expdecay(lags::AbstractVector{T}, acf::AbstractArray{T};
-                      dims::Int=ndims(acf); parallel::Bool=false) where {T <: Real}
-    slices = get_slices(acf, dims=dims)
+                      dims::Int=ndims(acf), parallel::Bool=false) where {T <: Real}
+    slices = collect.(get_slices(acf, dims=dims))
     f = x -> fit_expdecay(lags, vec(x))
     if parallel
-        return tmap(f, slices)
+        return stack_and_reshape(tmap(f, slices), dims=dims)
     else
-        return map(f, slices)
+        return stack_and_reshape(map(f, slices), dims=dims)
     end
 end
 
@@ -147,13 +147,13 @@ function fit_expdecay_3_parameters(lags::AbstractVector{T},
 end
 
 function fit_expdecay_3_parameters(lags::AbstractVector{T}, acf::AbstractArray{T};
-                                   dims::Int=ndims(acf); parallel=false) where {T <: Real}
-    slices = get_slices(acf, dims=dims)
+                                   dims::Int=ndims(acf), parallel=false) where {T <: Real}
+    slices = collect.(get_slices(acf, dims=dims))
     f = x -> fit_expdecay_3_parameters(lags, vec(x))
     if parallel
-        return tmap(f, slices)
+        return stack_and_reshape(tmap(f, slices), dims=dims)
     else
-        return map(f, slices)
+        return stack_and_reshape(map(f, slices), dims=dims)
     end
 end
 
@@ -191,12 +191,12 @@ end
 
 function acw50(lags::AbstractVector{T}, acf::AbstractArray{T};
                dims::Int=ndims(acf), parallel::Bool=false) where {T <: Real}
-    slices = get_slices(acf, dims=dims)
+    slices = collect.(get_slices(acf, dims=dims))
     f = x -> acw50(lags, vec(x))
     if parallel
-        return tmap(f, slices)
+        return stack_and_reshape(tmap(f, slices), dims=dims)
     else
-        return map(f, slices)
+        return stack_and_reshape(map(f, slices), dims=dims)
     end
 end
 
@@ -228,12 +228,12 @@ end
 
 function acw0(lags::AbstractVector{T}, acf::AbstractArray{S};
               dims::Int=ndims(acf), parallel::Bool=false) where {T <: Real, S <: Real}
-    slices = get_slices(acf, dims=dims)
+    slices = collect.(get_slices(acf, dims=dims))
     f = x -> acw0(lags, vec(x))
     if parallel
-        return tmap(f, slices)
+        return stack_and_reshape(tmap(f, slices), dims=dims)
     else
-        return map(f, slices)
+        return stack_and_reshape(map(f, slices), dims=dims)
     end
 end
 
@@ -265,12 +265,12 @@ end
 
 function acweuler(lags::AbstractVector{T}, acf::AbstractArray{S};
                   dims::Int=ndims(acf), parallel::Bool=false) where {T <: Real, S <: Real}
-    slices = get_slices(acf, dims=dims)
+    slices = collect.(get_slices(acf, dims=dims))
     f = x -> acweuler(lags, vec(x))
     if parallel
-        return tmap(f, slices)
+        return stack_and_reshape(tmap(f, slices), dims=dims)
     else
-        return map(f, slices)
+        return stack_and_reshape(map(f, slices), dims=dims)
     end
 end
 
@@ -295,12 +295,12 @@ end
 
 function acw_romberg(dt::Real, acf::AbstractArray{S};
                      dims::Int=ndims(acf), parallel::Bool=false) where {S <: Real}
-    slices = get_slices(acf, dims=dims)
[…]
