### 2025-06-19 94fa96e
implement parallel option in `acw`
