+++ src/core/acw.jl
@@ -55,7 +55,7 @@ possible_acwtypes = [:acw0, :acw50, :acweuler, :auc, :tau, :knee]
     acw(data, fs; acwtypes=possible_acwtypes, n_lags=nothing, freqlims=nothing, time=nothing, 
         dims=ndims(data), return_acf=true, return_psd=true, average_over_trials=false,
         trial_dims::Int=setdiff([1, 2], dims)[1], skip_zero_lag::Bool=false, max_peaks::Int=1, oscillation_peak::Bool=true,
-        allow_variable_exponent::Bool=false)
+        allow_variable_exponent::Bool=false, parallel::Bool=false)
 
 Compute various timescale measures for time series data. For detailed documentaion, see https://duodenum96.github.io/IntrinsicTimescales.jl/stable/acw/. 
 
@@ -82,6 +82,7 @@ Supported ACW types:
 - `max_peaks::Int=1`: Maximum number of oscillatory peaks to fit in spectral analysis
 - `oscillation_peak::Bool=true`: Whether to fit an oscillation peak in the spectral analysis
 - `allow_variable_exponent::Bool=false`: Whether to allow variable exponent in spectral fitting
+- `parallel::Bool=false`: Whether to use parallel computation
 
 # Returns
 - `ACWResults`: Structure containing computed ACW measures and intermediate results
@@ -101,7 +102,7 @@ Fields of the ACWResults structure:
 function acw(data, fs; acwtypes=possible_acwtypes, n_lags=nothing, freqlims=nothing, time=nothing, 
              dims=ndims(data), return_acf=true, return_psd=true, average_over_trials=false,
              trial_dims::Int=setdiff([1, 2], dims)[1], skip_zero_lag::Bool=false, max_peaks::Int=1, oscillation_peak::Bool=true,
-             allow_variable_exponent::Bool=false, constrained::Bool=false)
+             allow_variable_exponent::Bool=false, constrained::Bool=false, parallel::Bool=false)
 
     missingmask = ismissing.(data)
     if any(missingmask)
@@ -132,12 +133,12 @@ function acw(data, fs; acwtypes=possible_acwtypes, n_lags=nothing, freqlims=noth
 
     if any(in.(acf_acwtypes, [acwtypes]))
         if iscomplete
-            acf = comp_ac_fft(data; dims=dims)
+            acf = comp_ac_fft(data; dims=dims, parallel=parallel)
         else
             if isnothing(n_lags)
-                acf = comp_ac_time_missing(data; dims=dims)
+                acf = comp_ac_time_missing(data; dims=dims, parallel=parallel)
             else
-                acf = comp_ac_time_missing(data; dims=dims, n_lags=n_lags)
+                acf = comp_ac_time_missing(data; dims=dims, n_lags=n_lags, parallel=parallel)
             end
         end
 
@@ -148,10 +149,10 @@ function acw(data, fs; acwtypes=possible_acwtypes, n_lags=nothing, freqlims=noth
         lags_samples = 0.0:(size(data, dims)-1)
         lags = lags_samples * dt
 
-        acw0_sample = acw0(lags_samples, acf; dims=dims)
+        acw0_sample = acw0(lags_samples, acf; dims=dims, parallel=parallel)
         if any(in.(:acw0, [acwtypes]))
             acw0_idx = findfirst(acwtypes .== :acw0)
-            acw0_result = acw0(lags, acf; dims=dims)
+            acw0_result = acw0(lags, acf; dims=dims, parallel=parallel)
             if (acw0_result isa Vector) && (length(acw0_result) == 1)
                 result[acw0_idx] = acw0_result[1]
             else
@@ -161,7 +162,7 @@ function acw(data, fs; acwtypes=possible_acwtypes, n_lags=nothing, freqlims=noth
 
[…]
