### 2025-06-19 c68f61a
update docstrings to show parallel
