+++ src/utils/utils.jl
@@ -84,7 +84,7 @@ function residual_expdecay_3_parameters!(du, u, p)
 end
 
 """
-    fit_expdecay(lags, acf; dims=ndims(acf))
+    fit_expdecay(lags, acf; dims=ndims(acf), parallel=false)
 
 Fit exponential decay to autocorrelation function.
 
@@ -92,6 +92,7 @@ Fit exponential decay to autocorrelation function.
 - `lags::AbstractVector{T}`: Time lags
 - `acf::AbstractArray{T}`: Autocorrelation values
 - `dims::Int=ndims(acf)`: Dimension along which to fit
+- `parallel=false`: Whether to use parallel computation
 
 # Returns
 - Fitted timescale parameter(s)
@@ -114,14 +115,14 @@ function fit_expdecay(lags::AbstractVector{T}, acf::AbstractArray{T};
     slices = collect.(get_slices(acf, dims=dims))
     f = x -> fit_expdecay(lags, vec(x))
     if parallel
-        return stack_and_reshape(tmap(f, slices), dims=dims)
+        return tmap(f, slices)
     else
-        return stack_and_reshape(map(f, slices), dims=dims)
+        return map(f, slices)
     end
 end
 
 """
-    fit_expdecay_3_parameters(lags, acf)
+    fit_expdecay_3_parameters(lags, acf; parallel=false)
 
 Fit a 3-parameter exponential decay function to autocorrelation data ( A*(exp(-t/tau) + B) ). 
 Excludes lag 0 from fitting. 
@@ -129,6 +130,7 @@ Excludes lag 0 from fitting.
 # Arguments
 - `lags::AbstractVector{T}`: Time lags
 - `acf::AbstractVector{T}`: Autocorrelation values
+- `parallel=false`: Whether to use parallel computation
 
 # Returns
 - Fitted timescale parameter (tau)
@@ -151,9 +153,9 @@ function fit_expdecay_3_parameters(lags::AbstractVector{T}, acf::AbstractArray{T
     slices = collect.(get_slices(acf, dims=dims))
     f = x -> fit_expdecay_3_parameters(lags, vec(x))
     if parallel
-        return stack_and_reshape(tmap(f, slices), dims=dims)
+        return tmap(f, slices)
     else
-        return stack_and_reshape(map(f, slices), dims=dims)
+        return map(f, slices)
     end
 end
 
@@ -163,7 +165,7 @@ tau_from_knee(knee) = 1.0 ./ (2.0 .* pi .* knee)
 knee_from_tau(tau) = 1.0 ./ (2.0 .* pi .* tau)
 
 """
-    acw50(lags, acf; dims=ndims(acf))
+    acw50(lags, acf; dims=ndims(acf), parallel=false)
 
 Compute the ACW50 (autocorrelation width at 50%) along specified dimension.
 
@@ -171,6 +173,7 @@ Compute the ACW50 (autocorrelation width at 50%) along specified dimension.
 - `lags::AbstractVector{T}`: Vector of lag values
 - `acf::AbstractArray{T}`: Array of autocorrelation values
 - `dims::Int=ndims(acf)`: Dimension along which to compute ACW50
+- `parallel=false`: Whether to use parallel computation
 
 # Returns
 - First lag where autocorrelation falls below 0.5
@@ -194,14 +197,14 @@ function acw50(lags::AbstractVector{T}, acf::AbstractArray{T};
     slices = collect.(get_slices(acf, dims=dims))
     f = x -> acw50(lags, vec(x))
     if parallel
-        return stack_and_reshape(tmap(f, slices), dims=dims)
+        return tmap(f, slices)
     else
-        return stack_and_reshape(map(f, slices), dims=dims)
+        return map(f, slices)
     end
 end
 
 """
-    acw0(lags, acf; dims=ndims(acf))
[…]
