### 2025-06-19 0b114fa
add tests for parallel
