+++ test/test_acw.jl
@@ -83,13 +83,18 @@ using Random
         num_trials = 2
         data_2d = generate_ou_process(tau, 1.0, dt, duration, num_trials)
         result = acw(data_2d, fs, acwtypes=:acw0, dims=1).acw_results
+        result_parallel = acw(data_2d, fs, acwtypes=:acw0, dims=1, parallel=true).acw_results
         @test length(result) == size(data_2d, 2)
+        @test length(result_parallel) == size(data_2d, 2)
+        @test all(isapprox.(result, result_parallel))
 
         # Create 3D data
         data_3d = cat(data_2d, data_2d, dims=3)
         result = acw(data_3d, fs, acwtypes=:acw0, dims=1).acw_results
+        result_parallel = acw(data_3d, fs, acwtypes=:acw0, dims=1, parallel=true).acw_results
         @test size(result) == (size(data_3d, 2), size(data_3d, 3))
-
+        @test size(result_parallel) == (size(data_3d, 2), size(data_3d, 3))
+        @test all(isapprox.(result, result_parallel))
 
         # Test correctness of results along different dimensions using OU processes
         fs = 100
@@ -114,17 +119,20 @@ using Random
 
         # Test ACW50 along dimension 1
         result_test = acw(data_test, fs, acwtypes=:acw50, dims=2).acw_results
+        result_test_parallel = acw(data_test, fs, acwtypes=:acw50, dims=2, parallel=true).acw_results
         @test size(result_test) == (3, 1) # Check output shape
         
         # Test if ACW50 values increase with increasing tau
         # ACW50 should be proportional to tau: acw50 ≈ -tau * log(0.5)
         @test result_test[1] < result_test[2] < result_test[3]
+        @test all(isapprox.(result_test, result_test_parallel))
         @test isapprox(result_test[1], -tau1 * log(0.5), rtol=0.5)
         @test isapprox(result_test[2], -tau2 * log(0.5), rtol=0.5)
         @test isapprox(result_test[3], -tau3 * log(0.5), rtol=0.5)
 
         # Test tau estimation directly
         result_tau = acw(data_test, fs, acwtypes=:tau, dims=2).acw_results
+        result_tau_parallel = acw(data_test, fs, acwtypes=:tau, dims=2, parallel=true).acw_results
         @test size(result_tau) == (3, 1)
         @test isapprox(result_tau[1], tau1, rtol=1.0)
         @test isapprox(result_tau[2], tau2, rtol=0.5)
@@ -133,10 +141,12 @@ using Random
         # Test along dimension 2
         data_test_perm = permutedims(data_test, (2, 1, 3))
         result_test_dim2 = acw(data_test_perm, fs, acwtypes=:tau, dims=1).acw_results
+        result_test_dim2_parallel = acw(data_test_perm, fs, acwtypes=:tau, dims=1, parallel=true).acw_results
         @test size(result_test_dim2) == (3, 1)
         @test isapprox(result_test_dim2[1], tau1, rtol=1.0)
         @test isapprox(result_test_dim2[2], tau2, rtol=0.5)
         @test isapprox(result_test_dim2[3], tau3, rtol=0.5)
+        @test all(isapprox.(result_test_dim2, result_test_dim2_parallel))
     end
     
     @testset "Missing Data Handling" begin
@@ -206,14 +216,19 @@ using Random
             end
             
             result_2d = acw(data_2d_missing, fs, acwtypes=:tau, dims=2).acw_results
+            result_2d_parallel = acw(data_2d_missing, fs, acwtypes=:tau, dims=2, parallel=true).acw_results
             @test length(result_2d) == size(data_2d, 1)
             @test all(isfinite.(result_2d))
[…]
