### 2025-06-19 28615eb
bump version to 0.6
