+++ Project.toml
@@ -1,7 +1,7 @@
 name = "IntrinsicTimescales"
 uuid = "10d2f5c9-08bb-4a99-a156-10b2ac3b4091"
 authors = ["duodenum96 <catalyasir@gmail.com>"]
-version = "0.5.1"
+version = "0.6.0"
 
 [deps]
 AbstractFFTs = "621f4979-c628-5d54-868e-fcf4e3e8185c"
@@ -19,6 +19,7 @@ LombScargle = "fc60dff9-86e7-5f2f-a8a0-edeadbb75bd9"
 Missings = "e1d29d7a-bbdc-5cf2-9ac0-f12de2c33e28"
 NaNStatistics = "b946abbf-3ea7-4610-9019-9858bfdeaf2d"
 NonlinearSolve = "8913a72c-1f9b-4ce2-8d82-65094dcecaec"
+OhMyThreads = "67456a42-1dca-4109-a031-0a68de7e3ad5"
 Optimization = "7f7a1694-90dd-40f0-9382-eb1efda571ba"
 OptimizationOptimJL = "36348300-93cb-4f02-beb5-3c3902f8871e"
 Preferences = "21216c6a-2e73-6563-6e65-726566657250"
@@ -57,6 +58,7 @@ LombScargle = "1"
 Missings = "1"
 NaNStatistics = "0.6"
 NonlinearSolve = "4"
+OhMyThreads = "0.8.3"
 Optimization = "4"
 OptimizationOptimJL = "0.4"
 Plots = "1"
