### 2025-06-19 4149037
bump version to 0.6.0
