### 2025-07-01 d00aa9d
clarify `average_over_trials`
