### 2025-07-02 2a32bc2
add tutorials and reorganize docs
