+++ docs/src/home.md
@@ -58,9 +58,11 @@ The diagram below shows the rough organization of the package:
 
 ## Where to go from here?
 
-This documentation is divided in five parts. The fifth part [API](index.md) is an exhaustive list of functions and their signatures in the package. It is boring and most of the functions are not intended for end-user (you). A better place to start is the third and fourth parts, [Implementation](acw.md) and [Navigating the Forest of INT Metrics](tutorial/tutorial_1_acw.md). The implementation part documents [model-free](acw.md) and [simulation-based](simbasedinference.md) methods that are used in the package, with the full function signatures. This part should serve as the reference for you. The section longly titled [Navigating the Forest of INT Metrics](tutorial/tutorial_1_acw.md) intends to give practical considerations for picking which INT metric to use in which situation. If you are already familiar with calculating INTs and just want to start using the package, these two are right places. 
+This documentation is divided into three parts. [Explanation](practice/practice_intro.md) part is divided into theory and practice sections. The first section is [Practice](practice/practice_intro.md). It is usually easier to understand something after you do it, therefore, I placed the practice section before theory. In [Practice](practice/practice_intro.md), we carefully build our way towards estimating INTs by starting from the autocorrelation function and slowly proceeding to more and more advanced methods. If you never calculated INTs before, this is where you should start. The second part is [Theory](theory/theory.md). This part aims to delve into the history of INT research, what it means in the brain and what it is good for with a particular emphasis on theoretical research, summarizing the cutting edge in this frontier. It is especially useful for researchers working on INT itself. Right now, it is in construction. I will deploy it as soon as it is ready. 
 
-The remaining two parts are to understand the motivation to use various methods for calculating INTs and the motivation to calculate INTs (i.e., practice and theory). The first part is [Practice](practice/practice_intro.md). It is usually easier to understand something after you do it, therefore, I placed the practice section before theory. In [Practice](practice/practice_intro.md), we carefully build our way towards estimating INTs by starting from the autocorrelation function and slowly proceeding to more and more advanced methods. The second part is [Theory](theory/theory.md). This part delves into the history of INT research, what it means in the brain and what it is good for with a particular emphasis on theoretical research, summarizing the cutting edge in this frontier. It is especially useful for researchers working on INT itself. Right now, it is in construction. I will deploy it as soon as it is ready. 
+The third part is the [Reference](acw.md). This part contains the [Implementation](acw.md) and [API](index.md). The [API](index.md) is an exhaustive list of functions and their signatures in the package. It is boring and most of the functions are not intended for end-user (you). The implementation part documents [model-free](acw.md) and [simulation-based](simbasedinference.md) methods that are used in the package, with the full function signatures. This part should serve as the reference for the user. If you are already familiar with INTs and want to see how to use this package, you can start here. 
+
+Finally, the [Tutorial](tutorial/tutorial_1_acw.md) contains practical considerations for picking which INT metric to use in which situation, as well as [using the package with MNE-Python](tutorial/tutorial_2_mne.md) and [with FieldTrip](tutorial/tutorial_3_ft.md) packages. 
 
 ## Getting Help and Making Contributions
 
