### 2025-07-02 77949eb
fix equations
