### 2025-07-02 d436e88
bump version to 0.6.1
