+++ Project.toml
@@ -26,7 +26,6 @@ Preferences = "21216c6a-2e73-6563-6e65-726566657250"
 ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
-Revise = "295af30f-e4ad-537b-8983-00126c2a3abe"
 Romberg = "ff01fdde-158c-414c-948f-c94b4a236685"
 SciMLBase = "0bca4576-84f4-4d90-8ffe-ffa030f20462"
 SciMLSensitivity = "1ed8b502-d754-442c-8d5d-10ac956f44a1"
@@ -67,7 +66,6 @@ Preferences = "1"
 ProgressMeter = "1"
 Random = "1"
 Reexport = "1"
-Revise = "3.8.0"
 Romberg = "0.2"
 SciMLBase = "2"
 SciMLSensitivity = "7"
