### 2025-07-15 3b6e752
fix reproducibility issue and warning message in README and Quickstart
