### 2025-07-15 5752dee
typo fixes in practice_1_acf.md
