+++ docs/src/practice/practice_1_acf.md
@@ -6,7 +6,7 @@ To quantify the similarity between two things, we can use correlation. The highe
 
 ![](assets/practice_acf_1_drawio.svg)
 
-We took the time series x and shifted it forward in time by an amount `` \Delta``t. Then we need to take the correlation between them. To take a correlation between two things, you need to have equal number of data points in each of them. This is due to the definition of correlation, correlation is the average value of multiplication normalized by variance. You need to multiplicate corresponding data points. Take a look at the code example below. Throughout the documentation, there will be many code examples. I encourage you to run them on your computer and play around with them. Even better, take a pen and piece of paper and do the calculation below yourself. There is no better way to train intuition other than grinding your way through a calculation but I digress. Here is the code:
+We took the time series x and shifted it forward in time by an amount `` \Delta``t. Then we need to take the correlation between them. To take a correlation between two things, you need to have equal number of data points in each of them. This is due to the definition of correlation, correlation is the average value of multiplication normalized by variance. You need to multiply corresponding data points. Take a look at the code example below. Throughout the documentation, there will be many code examples. I encourage you to run them on your computer and play around with them. Even better, take a pen and piece of paper and do the calculation below yourself. There is no better way to train intuition other than grinding your way through a calculation but I digress. Here is the code:
 
 ```julia
 using Statistics # Import Statistics package for cor function
@@ -23,9 +23,9 @@ isapprox(correlation_x1_x2, cor(x1, x2)) # Compare with cor function from Statis
 
 This looks basic, but makes an important point. As you go forward in time, you need to match the time points in your time series and shifted version of it. In the figure above, the only usable part is the part indicated in black vertical lines. This means as we shift further in time, we have less time points at our disposal and our correlation results are less reliable. We will return back to this point later. 
 
-We took a time-series, shifted it by an amount ``\Delta``t, calculated the correlation and if the result is not zero, then we can say that the time series still haven't lost similarity to itself in ``\Delta``t amount of time. Take a moment to ponder about this sentence. We are insinuating that there is such a ``\Delta``t where the correlation is zero, or close to zero and this is the time it takes for a signal to lose similarity with itself. This is our INT. 
+We took a time-series, shifted it by an amount ``\Delta``t, calculated the correlation and if the result is not zero, then we can say that the time series still hasn't lost similarity to itself in ``\Delta``t amount of time. Take a moment to ponder about this sentence. We are insinuating that there is such a ``\Delta``t where the correlation is zero, or close to zero and this is the time it takes for a signal to lose similarity with itself. This is our INT. 
 
-Then a good strategy to calculate INT is simply calculating the correlation at various ``\Delta``t values and detecting which ``\Delta``t is the time where we lose correlation. Let's code this. We'll use the function [`generate_ou_process`](@ref) from the IntrinsicTimescales.jl package. This function simulates time series with a known timescale. I'll explain more about what it is doing in [Theory] section. For now, just know that this exists and is a good toy to play with. In IntrinsicTimescales.jl package, we have more optimized ways to do the operation I'll write below. I am doing this below explicitly and in detail so that we know exactly what we are doing when we compute these things. 
[…]
