### 2025-07-15 duodenum96/BayesianINT#41 @harisorgn
[joss] Confusing code block in One Time Scale Model doc
I think this field is quite confusing
https://github.com/duodenum96/IntrinsicTimescales.jl/blob/d41d0c6fe9eaae54b18ce1a02912ec35d48138b6/docs/src/one_timescale.md?plain=1#L14-L20
and I don't get its purpose there. People tend to copy/paste code from tutorials so this might cause errors. You could reference the API entry of this function if you want to show the function signature.
