### 2025-07-15 duodenum96/IntrinsicTimescales.jl#41 @harisorgn
I was mistaken, this is not a tutorial because you don't provide `data` etc. I got confused by all the code blocks. So I assume that you just want to showcase a workflow, it would be nice to do the same with an actual copy/paste-able code example. The field above is still confusing to me.
