### 2025-07-15 #8261 @harisorgn
I am happy for this package to go through once the issues I opened are fulfilled. Overall I stuck more to feedback on the documentation, as @duodenum96 has clearly put a lot of effort into it. Looks good overall!
