+++ docs/src/practice/practice_1_acf.md
@@ -78,7 +78,7 @@ vline!([acw_0], color=:red, label="ACW-0")
 
 So our work is done, right? We started with 1) the definition that INT is the time it takes for a time-series to lose its similarity with itself, 2) operationalized similarity with correlation, 3) operationalized similarity with itself as correlation with itself shifted some time lags and 4) identified the INT as the number of time lags required to lose similarity. There is one problem. Remember the problem of number of time points we talked about above. As we go further in lags, we have less and less number of data points to calculate the correlation, the portion inside vertical black lines is getting smaller and smaller. If we do not have enough number of data points to calculate ACW-0, then we will get a noisy estimate. 
 
-Let's try to see how big of a problem this is. Below, we will simulate the time-series again and again and overlay plots of ACFs. In a different panel, we'll do a histogram of ACW-0 values. To calculate ACF, we will use the function [`comp_ac_fft`](@ref) from  IntrinsicTimescales.jl package. This function is faster and uses a different technique to calculate ACF which I'll explain in the [next section]. For now, it should suffice to know that it takes the data and optionally the number of lags we want as input and gives back the ACF. If number of lags is not specified, it goes through all possible lags. To get the ACW-0 from the ACF, we'll use [`acw0`](../acw.md) function which takes lags and ACF as input and gives ACW-0 value. 
+Let's try to see how big of a problem this is. Below, we will simulate the time-series again and again and overlay plots of ACFs. In a different panel, we'll do a histogram of ACW-0 values. To calculate ACF, we will use the function [`comp_ac_fft`](@ref) from  IntrinsicTimescales.jl package. This function is faster and uses a different technique to calculate ACF which I'll explain in the [next section](practice_2_acw.md). For now, it should suffice to know that it takes the data and optionally the number of lags we want as input and gives back the ACF. If number of lags is not specified, it goes through all possible lags. To get the ACW-0 from the ACF, we'll use [`acw0`](../acw.md) function which takes lags and ACF as input and gives ACW-0 value. 
 
 ```julia
 acw0_results = [] # Initialize empty vectors to hold the results
