### 2025-07-15 441d88a
fix sentence that ends abruptly, add link to the next section
