### 2025-07-15 b202173
fix typo of `\omega` -> ``\omega``
