### 2025-07-17 7b70e80
make plots reproducible for practice_2
