+++ docs/src/practice/practice_2_acw.md
@@ -6,13 +6,13 @@ First, some nomenclature. When we make an analysis on the data, for example, cal
 
 ![](assets/practice_2_estimator.drawio.svg)
 
-Our first note about the noise of estimators was the finiteness of the data. We noted that as we go along further lags, we have less data points at our hand to calculate the correlation values, making the estimate noisier. A first response to the problem is to use a different cutoff. Instead of waiting the autocorrelation function to reach exactly to 0 thus completely losing the similarity, we can cut it off when it reaches 0.5 and say losing half of the similarity. After all, a time-series with a longer timescale should take longer to lose half of it. This method is called ACW-50. It is  older than ACW-0. To my knowledge, used first in [Honey et al., 2012](https://pubmed.ncbi.nlm.nih.gov/23083743/). This was a time when the phrase intrinsic neural timescale had not been established. The term at that time was temporal receptive windows (TRW). I will discuss the evolution of the term more in the [Theory](../theory/theory.md) section. For now, we will make simulations from two processes with different timescales and see how well we can distinguish their INTs using ACW-50 versus ACW-0. To quickly get many simulations with the same timescale, I will set num\_trials to 1000 in the function [`generate_ou_process`](@ref). 
+Our first note about the noise of estimators was the finiteness of the data. We noted that as we go along further lags, we have less data points at our hand to calculate the correlation values, making the estimate noisier. A first response to the problem is to use a different cutoff. Instead of waiting the autocorrelation function to reach exactly to 0 thus completely losing the similarity, we can cut it off when it reaches 0.5 and say losing half of the similarity. After all, a time-series with a longer timescale should take longer to lose half of it. This method is called ACW-50. It is  older than ACW-0. To my knowledge, used first in [Honey et al., 2012](https://pubmed.ncbi.nlm.nih.gov/23083743/). This was a time when the phrase intrinsic neural timescale had not been established. The term at that time was temporal receptive windows (TRW). I will discuss the evolution of the term more in the [Theory](../theory/theory.md) section. For now, we will make simulations from two processes with different timescales and see how well we can distinguish their INTs using ACW-50 versus ACW-0. To quickly get many simulations with the same timescale, I will set num\_trials to 1000 in the function [`generate_ou_process`](@ref). Note that I am also setting `rng` and `deq_seed` arguments, these are for reproducibility. `rng` here handles the reproducibility in the initial conditions (e.g. the starting value of simulations) and `deq_seed` makes sure the rest of the simulations are reproducible. 
 
 ```julia
 using IntrinsicTimescales # import IntrinsicTimescales package
 using Random 
 using Plots # to plot the results
-Random.seed!(1) # for replicability
+Random.seed!(1)
 
 timescale_1 = 1.0
 timescale_2 = 3.0
@@ -23,7 +23,7 @@ num_trials = 1000 # Number of trials
 
 data_1 = generate_ou_process(timescale_1, sd, dt, duration, num_trials)
 data_2 = generate_ou_process(timescale_2, sd, dt, duration, num_trials)
-println(size(data_1)) # == 30, 1000: 30 trials and 10000 time points
[…]
