### 2025-07-17 de4f31e
reproducibility for practice_3
