### 2025-07-17 1127e14
remove rng and deq_seed in generate_ou_process
