### 2025-07-17 c1cffa5
reproducibility for practice_5
