### 2025-07-17 5d7cdc1
bump version to 0.6.2
