### 2025-07-18 66667a5
fix bug related to freqlims that was making informed_priors usable for summary_statistic=:psd
