### 2025-07-18 11a1460
correct fit -> int_fit
