### 2025-07-18 7eebc91
refactor fit->int_fit and give example, fix bugs
