### 2025-07-18 84ca826
copy-paste friendly `acw.md`
