### 2025-07-18 cad4ab3
copy-paste friendly acw.md
