### 2025-07-18 duodenum96/IntrinsicTimescales.jl#41 @harisorgn
This change clarifies things more, thanks. My suggestion would be to just link to the API page instead of showing the signature here. When it comes to function signatures, that would be my go-to, and the standard in Jula package docs in general I'd say, as API should work like a lookup table for all exported functions with details on input arguments etc. But this is only a stylistic suggestion and a matter of preference, feel free to ignore it if you want to. I am happy with how you addressed it already.
