### 2025-07-28 #8261 @duodenum96
Hello @britta-wstnr, currently there are no open issues related to the review. There is one issue that I intend to revisit (https://github.com/duodenum96/IntrinsicTimescales.jl/issues/41) but it is also closed and the main problem resolved.
