### 2025-08-12 2a02d85
clarify docstrings of *_model
