### 2025-08-12 975ad78
bump version to 0.6.3
