### 2025-08-30 8822d71
update paper.md and paper.bib
