+++ paper/paper.bib
@@ -315,7 +315,7 @@ Publisher: Nature Publishing Group},
 }
 
 @article{beaumont_adaptive_2009,
-	title = {Adaptive approximate Bayesian computation},
+	title = {Adaptive approximate {Bayesian} computation},
 	volume = {96},
 	issn = {0006-3444},
 	url = {https://www.jstor.org/stable/27798882},
