+++ paper/paper.md
@@ -22,19 +22,19 @@ bibliography: paper.bib
 
 # Summary
 
-IntrinsicTimescales.jl is a Julia package to perform estimation of intrinsic neural timescales (INTs). INTs are defined as the time window in which prior information from an ongoing stimulus can affect the processing of newly arriving information[@hasson_hierarchical_2015; @hasson_hierarchy_2008; @wolff_intrinsic_2022; @golesorkhi_brain_2021]. INTs are estimated either from the autocorrelation function (ACF) or the power spectral density (PSD) of time-series data [@honey_slow_2012; @gao_neuronal_2020]. In addition to the model-free estimates of INTs, IntrinsicTimescales.jl offers implementations of novel techniques of timescale estimation, estimating parameters of an Ornstein-Uhlenbeck process with adaptive approximate Bayesian computation (aABC) [@beaumont_adaptive_2009; @zeraati_flexible_2022] and automatic differentiation variational inference (ADVI) [@kucukelbir_automatic_2017]. 
+`IntrinsicTimescales.jl` is a Julia package to perform estimation of intrinsic neural timescales (INTs). INTs are defined as the time window in which prior information from an ongoing stimulus can affect the processing of newly arriving information[@hasson_hierarchical_2015; @hasson_hierarchy_2008; @wolff_intrinsic_2022; @golesorkhi_brain_2021]. INTs are estimated either from the autocorrelation function (ACF) or the power spectral density (PSD) of time-series data [@honey_slow_2012; @gao_neuronal_2020]. In addition to the model-free estimates of INTs, `IntrinsicTimescales.jl` offers implementations of novel techniques of timescale estimation, estimating parameters of an Ornstein-Uhlenbeck process with adaptive approximate Bayesian computation (aABC) [@beaumont_adaptive_2009; @zeraati_flexible_2022] and automatic differentiation variational inference (ADVI) [@kucukelbir_automatic_2017]. 
 
 # Statement of Need
 
 Intrinsic neural timescales (INTs) were found to be an important metric to probe brain dynamics and function. On the neuroscientific side, INTs were found to follow the large-scale gradients in the cortex ranging from uni- to transmodal areas, including local and long-range excitation [@murray_hierarchy_2014; @wang_macroscopic_2020] and proxies of myelination [@ito_cortical_2020; @wang_macroscopic_2020]. From a cognitive science perspective, INTs were found to be related to reward [@murray_hierarchy_2014], behavior [@zeraati_intrinsic_2023; @catal_flexibility_2024], self [@wolman_intrinsic_2023], and consciousness [@zilio_are_2021], among others. Proper estimation of INTs is crucial to make sure the estimates are not affected by limited data, missingness of the data, or oscillatory artifacts. While several methods exist for estimating INTs, there is a lack of standardized, open-source tools that implement both traditional model-free approaches and modern Bayesian estimation techniques. Existing software solutions are often limited to specific estimation methods, lack proper uncertainty quantification, or are not optimized for large-scale neuroimaging data.
 
-IntrinsicTimescales.jl addresses these limitations by providing a comprehensive, high-performance toolbox for INT estimation. The package implements both established model-free methods and novel Bayesian approaches, allowing researchers to compare and validate results across different methodologies with a simple API. Its implementation in Julia ensures computational efficiency, crucial for analyzing large neuroimaging datasets. The package's modular design facilitates easy extension and integration with existing neuroimaging workflows, while its rigorous testing and documentation make it accessible to researchers across different levels of programming expertise.
[…]
