### 2025-09-14 9c0fa40
bump version to 0.6.4 (JOSS release)
