### 2025-09-14 #8261 @duodenum96
Hello @britta-wstnr ,

I confirm going through the checklist. Here is the DOI for archived repository in Zenodo:

https://doi.org/10.5281/zenodo.17117276

The final version so far is v0.6.4.

Thanks again for your time!
