### 2025-09-16 efe83dd
add abcdez extension and tests
