+++ test/test_abcdez.jl
@@ -1,0 +1,285 @@
+using Test
+using Distributions
+using IntrinsicTimescales
+using Statistics
+using Random
+
+
+@testset "ABCdeZ Extension Tests" begin
+    
+    @testset "Parameter Dictionary Functions" begin
+        
+        @testset "get_param_dict_abcdemc" begin
+            param_dict = IntrinsicTimescales.get_param_dict_abcdemc()
+            
+            # Test that it returns a dictionary
+            @test param_dict isa Dict
+            
+            # Test all required keys are present
+            required_keys = [:nparticles, :generations, :verbose, :rng, :parallel]
+            for key in required_keys
+                @test haskey(param_dict, key)
+            end
+            
+            # Test specific values
+            @test param_dict[:nparticles] == 50
+            @test param_dict[:generations] == 20
+            @test param_dict[:verbose] == true
+            @test param_dict[:rng] == Random.GLOBAL_RNG
+            @test param_dict[:parallel] == true
+            
+            # Test types
+            @test param_dict[:nparticles] isa Integer
+            @test param_dict[:generations] isa Integer
+            @test param_dict[:verbose] isa Bool
+            @test param_dict[:parallel] isa Bool
+        end
+        
+        @testset "get_param_dict_abcdesmc" begin
+            param_dict = IntrinsicTimescales.get_param_dict_abcdesmc()
+            
+            # Test that it returns a dictionary
+            @test param_dict isa Dict
+            
+            # Test all required keys are present
+            required_keys = [:nparticles, :α, :δess, :nsims_max, :Kmcmc, :Kmcmc_min, 
+                           :ABCk, :facc_min, :facc_tune, :verbose, :verboseout, :rng, :parallel]
+            for key in required_keys
+                @test haskey(param_dict, key)
+            end
+            
+            # Test specific values
+            @test param_dict[:nparticles] == 100
+            @test param_dict[:α] == 0.95
+            @test param_dict[:δess] == 0.5
+            @test param_dict[:nsims_max] == 10^7
+            @test param_dict[:Kmcmc] == 3
+            @test param_dict[:Kmcmc_min] == 1.0
+            @test param_dict[:facc_min] == 0.15
+            @test param_dict[:facc_tune] == 0.975
+            @test param_dict[:verbose] == true
+            @test param_dict[:verboseout] == true
+            @test param_dict[:rng] == Random.GLOBAL_RNG
+            @test param_dict[:parallel] == false
+            
+            # Test types
+            @test param_dict[:nparticles] isa Integer
+            @test param_dict[:α] isa Real
+            @test param_dict[:δess] isa Real
+            @test param_dict[:nsims_max] isa Integer
+            @test param_dict[:Kmcmc] isa Integer
+            @test param_dict[:Kmcmc_min] isa Real
+            @test param_dict[:facc_min] isa Real
+            @test param_dict[:facc_tune] isa Real
+            @test param_dict[:verbose] isa Bool
+            @test param_dict[:verboseout] isa Bool
+            @test param_dict[:parallel] isa Bool
+            
+            # Test value ranges
+            @test 0 < param_dict[:α] < 1
+            @test 0 < param_dict[:δess] < 1
+            @test param_dict[:nsims_max] > 0
+            @test param_dict[:Kmcmc] > 0
+            @test param_dict[:Kmcmc_min] > 0
+            @test 0 < param_dict[:facc_min] < 1
+            @test 0 < param_dict[:facc_tune] < 1
+        end
+    end
+    
+    @testset "ABCdeZ Inference Functions" begin
+        # Create a simple test model for testing inference functions
+        Random.seed!(123)
+        
+        # Generate test data - simple OU process
+        n_time = 100
+        time = collect(0.0:0.01:(n_time-1)*0.01)
+        true_tau = 50.0
[…]
