+++ src/core/abcdez.jl
@@ -1,0 +1,10 @@
+module ABCDEZ
+
+export get_param_dict_abcdemc, get_param_dict_abcdesmc, 
+    abcdez_inference
+
+function get_param_dict_abcdemc end
+function get_param_dict_abcdesmc end
+function abcdez_inference end
+
+end
