### 2025-09-16 e9d9c8b
get posterior and fix tests, move abcdez to hard dependencies for now
