### 2025-10-17 #8261 @duodenum96
Thanks so much to everyone involved! Both the paper and the software is in much better shape thanks to your comments.

PS. @britta-wstnr, already signed up as a reviewer.
